% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actstudio-package.R
\docType{package}
\name{actstudio-package}
\alias{actstudio}
\alias{actstudio-package}
\title{actstudio: A Set of Tools for Actigraphy/Actimetry Data Manipulation}
\description{
Open source package of applications for processing, analyzing 
    and visualizing actigraphy/actimetric data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gipsousp/actstudio}
  \item Report bugs at \url{https://github.com/gipsousp/actstudio/issues}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID})

Authors:
\itemize{
  \item Bruno Furlan \email{brfrlan@gmail.com} (\href{https://orcid.org/0000-0003-1654-0899}{ORCID})
  \item Cássio Serrano \email{cassio.serrano@usp.br} (\href{https://orcid.org/0000-0001-7149-1829}{ORCID})
  \item Vinícius Matias \email{viniciusmatias@usp.br} (\href{https://orcid.org/0000-0002-2047-5420}{ORCID})
  \item Ana Amélia Benedito-Silva \email{aamelia@usp.br} [expert, scientific advisor]
  \item Mario Pedrazzoli \email{pedrazzo@usp.br} (\href{https://orcid.org/0000-0002-5257-591X}{ORCID}) [expert, scientific advisor]
}

Other contributors:
\itemize{
  \item GIPSO [copyright holder, funder]
  \item USP [funder]
}

}
\keyword{internal}
