% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acttrust.R
\docType{data}
\name{acttrust}
\alias{acttrust}
\title{An ActTrust 1 actigraphy record}
\format{
A \code{\link[tsibble:tsibble]{tsibble}} with 17 columns and 1,441 rows:

\describe{
\item{timestamp}{
Date and time of the recorded events.
\cr \cr
R class: \code{\link[=as.POSIXct]{POSIXct}}.
}

\item{pim}{
Wrist activity in Proportional Integral Mode (PIM).
\cr \cr
R class: \code{\link[=numeric]{numeric}}.
}

\item{tat}{
Wrist activity in Time Above Threshold (TAT) mode.
\cr \cr
R class: \code{\link[=numeric]{numeric}}.
}

\item{zcm}{
Wrist activity in Zero Crossing Mode (ZCM).
\cr \cr
R class: \code{\link[=numeric]{numeric}}.
}

\item{orientation}{
Device orientation.
\cr \cr
R class: \code{\link[=numeric]{numeric}}.
}

\item{wrist_temperature}{
Wrist temperature in degrees \code{Celsius}.
\cr \cr
R class: \code{\link[=numeric]{numeric}}.
}

\item{external_temperature}{
External temperature in degrees \code{Celsius} (unreliable data).
\cr \cr
R class: \code{\link[=numeric]{numeric}}.
}

\item{light}{
Light intensity in \code{lux}.
\cr \cr
R class: \code{\link[=numeric]{numeric}}.
}

\item{ambient_light}{
Ambient light intensity in \verb{microwatts/square meter}.
\cr \cr
R class: \code{\link[=numeric]{numeric}}.
}

\item{red_light}{
Red light intensity in \verb{microwatts/square centimeter}.
\cr \cr
R class: \code{\link[=numeric]{numeric}}.
}

\item{green_light}{
Green light intensity in \verb{microwatts/square centimeter}.
\cr \cr
R class: \code{\link[=numeric]{numeric}}.
}

\item{blue_light}{
Blue light intensity in \verb{microwatts/square centimeter}.
\cr \cr
R class: \code{\link[=numeric]{numeric}}.
}

\item{ir_light}{
Infrared light intensity in \verb{microwatts/square centimeter}.
\cr \cr
R class: \code{\link[=numeric]{numeric}}.
}

\item{uva_light}{
Ultraviolet A light intensity in \verb{microwatts/square centimeter}.
\cr \cr
R class: \code{\link[=numeric]{numeric}}.
}

\item{uvb_light}{
Ultraviolet B light intensity in \verb{microwatts/square centimeter}.
\cr \cr
R class: \code{\link[=numeric]{numeric}}.
}

\item{event}{
A binary number indicating the moments when the subject pressed the event
button: \code{0} when the button is not pressed and \code{1} when the button is
pressed. This is commonly used to indicate when the subject goes to sleep
or wakes up.
\cr \cr
R class: \code{\link[=numeric]{numeric}}.
}

\item{state}{
An integer number indicating the subject state: \code{0} for awake, \code{1} for
sleeping, \code{2} for resting, \code{4} for offwrist (device removal), \code{6} for
editable state 1, \code{7} for editable state 2, and \code{8} for editable state 3.
These states can be assigned by a specialist or by a validated inference
algorithm.
\cr \cr
R class: \code{\link[=numeric]{numeric}}.
}
}
}
\source{
Created by \href{https://linktr.ee/danielvartan}{Daniel Vartanian}
(package author).
}
\usage{
acttrust
}
\description{
A small sample (1 day length) of an \href{https://condorinst.com/en/}{ActTrust}
(model 1) actigraphy record.

ActTrust is a trademark of
\href{https://condorinst.com/en/}{Condor Instruments Ltda}.

\strong{Note}: This dataset is based on older methodologies, some of which have
since been improved or replaced. For up-to-date recommendations on sleep
scoring, please refer to the Brazilian Consensus on Actigraphy
(\href{https://doi.org/10.5281/zenodo.15377336}{Pedrazzoli & Gonçalves, 2021}).
}
\details{
This dataset was created by one of the \code{actverse} package authors and is free
for all kinds of use. You can download the full dataset at:
\url{https://zenodo.org/record/4898822}
\subsection{Raw data}{

The raw data for this dataset can be found by running the following code:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{[`get_raw_data("acttrust.txt")`][get_raw_data()].
}\if{html}{\out{</div>}}
}

\subsection{Data munging}{

This dataset was processed with the default settings of
\code{\link[=read_acttrust]{read_acttrust()}}. Details about the scoring
process can be found in the Metadata section.
}

\subsection{Variable naming}{

Variable names follow the
\href{https://style.tidyverse.org/}{tidyverse style guide},
using lowercase letters and underscores to ensure clarity and consistency.
}

\subsection{Variable classes}{

R supports a \href{https://cran.r-project.org/view=TimeSeries}{wide variety} of
time series data formats. We use \code{\link[tsibble:tsibble]{tsibble}} as it aligns
well with \href{https://tidyverse.tidyverse.org/articles/manifesto.html}{tidyverse principles} and
offers a modern, consistent approach to time series data.

To facilitate interoperability with other standards, all variables are stored
as \code{\link[=numeric]{numeric}} vectors except for the index, which is of class
\code{\link[=as.POSIXct]{POSIXct}}. This design choice ensures that data from
\code{actverse} can be easily converted to other formats if needed.
}
}
\section{Metadata}{

\subsection{Subject}{
\itemize{
\item Age: \code{35}
\item Gender: Male
\item Height: \code{180} cm
\item Weight: \code{85} kg
\item BMI: \code{~26.235} (overweight)
}
}

\subsection{Record}{
\itemize{
\item Body part: Left wrist (non-dominant wrist)
\item Start: \verb{2021-04-24 04:14:25}
\item End: \verb{2021-05-31 03:39:06}
\item Days: \code{37}
\item Epoch (sample interval): \code{60} seconds
\item Mode: PIM, TAT, and ZCM
\item TAT Threshold: \code{1024}
\item Location: São Paulo, SP, Brazil
\item UTC: \code{-3}
\item DST: \code{FALSE}
\item Light phase: \code{06:23:59} - \code{17:44:58}
\item Dark phase: \code{17:44:59} - \code{06:23:58}
}
}

\subsection{Actigraph}{
\itemize{
\item Brand: \href{https://condorinst.com/en/}{Condor Instruments}
\item Model: \href{https://condorinst.com/en/}{ActTrust 1}
\item Hardware version: \code{3.1}
\item Firmware version: \code{3.7}
\item Manual: \verb{1.0.13} (PT-BR)
\item Software: ActStudio Alpha 1.0.17 (Win | Mac)
}
}

\subsection{Scoring algorithm settings}{
\itemize{
\item Condor Instruments algorithm version 1.0.17.
\item The sleep diary data was not used for scoring.
}
}

\subsection{Scoring review}{
\itemize{
\item Offwrist periods were indicated from sudden drops in wrist temperature.
}
\itemize{
\item Micro sleep states (isolated states lasting less than \verb{5 minutes}) were
removed.
}
\itemize{
\item Minor adjustments were made on the sleep onset and sleep offset, taking
into consideration the event button and sleep diary records.
}
}

\subsection{Sleep statistics settings}{
\itemize{
\item Multiple main sleep periods: \code{TRUE}.
\item Main sleep period time of day: \verb{Day or Night}.
\item Minimum duration for main sleep periods: \code{01:00:00}.
\item Awake period limit: \code{02:00:00}.
}
}
}

\references{
Pedrazzoli, M., Gonçalves, B. da S. B., Benedito-Silva, A. A., Toscanini, A.
C., Souza, A. P. P. B. de, Conway, B. A., Vartanian, D., Santos, E. H. R.,
Pires, M. L. N., & Leocadio-Miguel, M. A. (2021). \emph{Consenso brasileiro de
actigrafia}. Segmento Farma Editores. \doi{10.5281/zenodo.15377336}
}
\concept{datasets}
\keyword{datasets}
