% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actverse-package.R
\docType{package}
\name{actverse-package}
\alias{actverse}
\alias{actverse-package}
\title{actverse: Process, Visualize, and Analyze Actigraphy Data}
\description{
A comprehensive toolkit to process, analyze and visualize actigraphy data. Designed to support sleep and chronobiology researchers by streamlining workflows and enhancing reproducibility in actigraphy research.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://danielvartan.github.io/actverse/}
  \item \url{https://github.com/danielvartan/actverse/}
  \item Report bugs at \url{https://github.com/danielvartan/actverse/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [conceptor, copyright holder]

Authors:
\itemize{
  \item Vinicius Alves Matias \email{viniciusmatias@usp.br} (\href{https://orcid.org/0000-0002-2047-5420}{ORCID})
  \item Cassio Almeida Mattos Serrano \email{cassio.serrano@usp.br} (\href{https://orcid.org/0000-0001-7149-1829}{ORCID})
  \item Ana Amélia Benedito-Silva \email{aamelia@usp.br} (\href{https://orcid.org/0000-0003-0667-9504}{ORCID}) [scientific advisor]
}

Other contributors:
\itemize{
  \item Bruno Ramalho Furlan \email{brfrlan@gmail.com} (\href{https://orcid.org/0000-0003-1654-0899}{ORCID}) [contributor]
}

}
\keyword{internal}
