% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_index.R
\name{aggregate_index}
\alias{aggregate_index}
\title{Aggregate the index of a \code{tsibble}}
\usage{
aggregate_index(data, unit, fun = NULL, week_start = 1)
}
\arguments{
\item{data}{A \code{\link[tsibble:tsibble]{tsibble}} object.}

\item{unit}{A string indicating at which time unit the index must be
aggregated. Valid values are: \verb{“seconds”}, \verb{“minutes”}, \verb{“hours”},
\verb{“days”}, \verb{“weeks”}, \verb{“months”}, \verb{“quarters”}, and \verb{“years”}) (default:
\code{"minutes"}).}

\item{fun}{(optional) A \code{\link[base:is.function]{function}} to be applied to
each measure variable of \code{data}. If \code{NULL}, \code{aggregate_index()} will apply
its default function (see the Details section to learn more)
(default: \code{NULL}).}

\item{week_start}{(optional) An integer number indicating the day on which
the week starts (\code{1} for Monday and \code{7} for \code{Sunday}). This is only used
when \code{unit} is set to \code{"weeks"} (default: \code{1}).}
}
\value{
A \code{\link[tsibble:tsibble]{tsibble}} object with the same columns as
\code{data}, but with the index aggregated to the specified time unit.
}
\description{
\code{aggregate_index()} allows you to aggregate the index of a
\code{\link[tsibble:tsibble]{tsibble}} object by applying a specific function to
its measured variables.
}
\details{
\code{aggregate_index()} is designed to simplify the regularization of
\code{\link[tsibble:tsibble]{tsibble}} objects by aggregating their index to a
specified time unit. For more advanced or customized aggregation, consider
using the \code{\link[tsibble:index-by]{index_by()}} function from the
\href{https://tsibble.tidyverts.org/}{\code{tsibble}} package, which offers greater
flexibility and control.
\subsection{Default function}{

If the \code{fun} argument is \code{NULL}, \code{aggregate_index()} will use the following
function to transform each measured variable:

\if{html}{\out{<div class="sourceCode">}}\preformatted{function(x) \{
    if (is.numeric(x) && !all(nchar(x) == 1, na.rm = TRUE)) \{
        mean(x, na.rm = TRUE)
    \} else \{
        y <- x[which(!is.na(x))]
        unique <- unique(y)
        unique[which.max(tabulate(match(y, unique)))]
    \}
\}
}\if{html}{\out{</div>}}

This function computes the mean for numeric variables and assigns the most
frequent value (mode) for non-numeric variables. If no mode is found, it
returns the first non-missing value of \code{x}.
}
}
\examples{
acttrust

aggregate_index(acttrust, unit = "hour")

aggregate_index(acttrust, unit = "day")

aggregate_index(acttrust, unit = "week")

aggregate_index(acttrust, unit = "month")

aggregate_index(acttrust, unit = "quarter")

aggregate_index(acttrust, unit = "year")
}
\seealso{
Other utility functions: 
\code{\link{find_epoch}()},
\code{\link{get_raw_data}()}
}
\concept{utility functions}
