% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cole_kripke.R
\name{cole_kripke}
\alias{cole_kripke}
\title{Classifies activity using the Cole-Kripke algorithm}
\usage{
cole_kripke(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
The same data with two columns: STAGE and AWAKE.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\strong{UNDER DEVELOPMENT}
}
\examples{
\dontrun{
data <- readr::read_csv2("./inst/extdata/test_log.txt", skip = 25)
data <- cole_kripke(data)
data$timestamp <- lubridate::dmy_hms(paste(data$DATE, data$TIME))
ggplot2::qplot(timestamp, AWAKE, data = data, geom = "area",
               xlab = "Timestamp", ylab = "State")
}
}
