% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epoch_change.R
\name{epoch_change}
\alias{epoch_change}
\title{Change the epoch length of an actimetry dataset}
\usage{
epoch_change(
  data,
  new_epoch,
  method = "sum",
  timestamp = "timestamp",
  variable = NULL
)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{new_epoch}{A integer number indicating the number of seconds of the
new epoch.

Note that when the new epoch is smaller than the original, the value
must be an integer divisor of the original epoch. (e.g. if the original
epoch is 60, the possible values are 30 (60 \%\% 30 = 0), 20,
15, 12, 10, 6, 5, 4, 3 , 2 and 1)}

\item{method}{A string indicating the method that must be used for data
summarization.

Note that some methods are not valid when dealing with a lower or greater
epoch value than the original. You can use the \code{\link[=epoch]{epoch()}} function to know
the epoch value of your data.

This variable accepts only the following values.
\itemize{
\item \code{"divide"}: divide and equally distribute the value from the original
epoch into the new epochs created (only when new epoch \code{<} original epoch)
\item \code{"same"}: assign the same value from the original epoch to the new epochs
created (only when new epoch \code{<} original epoch)
\item \code{"first"}: use the first value from original epochs grouped in the
new epoch interval (only when new epoch \code{>} original epoch)
\item \code{"last"}: use the last value from original epochs grouped in the
new epoch interval (only when new epoch \code{>} original epoch)
(only when new epoch \code{>} original epoch)
\item \code{"sum"}: sum all the values in the new epoch interval
(only when new epoch \code{>} original epoch)
\item \code{"mean"}: compute the mean value in the new epoch interval (only when new
epoch \code{>}original epoch)
\item \code{"median"}: compute the median value in the new epoch interval (only
when new epoch \code{>}original epoch)
}}

\item{timestamp}{A string indicating the name of the timestamp
variable (Note: the timestamp variable must be class \code{POSIXct}).}

\item{variable}{a string indicating a specific variable in \code{data} (optional).
If not assigned, its required that the data conforms to the proposed data
structure of the \code{\link[=tidy_data]{tidy_data()}} function.

When using a tidy dataset, the method will be used only for the
activity and light exposure variables. The variables \code{orientation},
\code{body_temperature}, and \code{external_temperature} will be summarized by
the \code{mean} method. If there's a \code{TRUE} value for the \code{event} variable
in the time interval assessed, the final value will be \code{TRUE}. For the
\code{state} variable, the first value of the time interval assessed will be used.}
}
\value{
A tibble.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\strong{UNDER DEVOLOPMENT}

Some computations may require a specific epoch length. This function change
the epoch length using a method defined in the \code{method} variable. Be aware
that this kind of data transformation can produce misleading results.

Note that some methods are not valid when dealing with a lower or greater
epoch value than the original. Use \code{\link[=epoch]{epoch()}} to know the
epoch value of your data.
}
\note{
To do list:
\itemize{
\item Optimize the matching process
}
}
\section{Representations}{


The representations below can help you visualize the right kind of
transformation that you need.
\itemize{
\item Example of an epoch length
}\preformatted{       00:00:00   00:01:00   00:02:00
----------|----------|----------|----------
          |----------|
             epoch
}
\itemize{
\item Example for when New epoch \code{<} Original epoch
}\preformatted{       00:00:00   00:00:30   00:01:00
----------|----------|----------|----------
          |---------------------|
               original epoch
          |----------|
           new epoch
}
\itemize{
\item Example for when New epoch \code{>} Original epoch
}\preformatted{       00:00:00   00:00:30   00:01:00
----------|----------|----------|----------
          |----------|
         original epoch
          |---------------------|
                  new epoch
}
}

\examples{
\dontrun{
epoch_change(test_log)}
}
