% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_epoch.R
\name{find_epoch}
\alias{find_epoch}
\title{Find the epochs/periodicities of a time series}
\usage{
find_epoch(data, threshold = 0.9)
}
\arguments{
\item{data}{A \code{\link[zoo:zoo]{zoo}} or an \code{\link[xts:xts]{xts}} object with
a \code{\link[=as.Date]{Date}} or \link[=as.POSIXct]{POSIXt} vector as index.}

\item{threshold}{A number, from 0 to 1, indicating the minimum proportion of
that an epoch must have to be considered valid. \code{threshold = 1} means
strict regularity.}
}
\value{
a \code{\link[=list]{list}} object with the following elements:
\itemize{
\item \code{best_match}: A number indicating the epoch/periodicity with greater
prevalence above \code{threshold} in seconds. If none is find, \code{best_match} value
will be \code{as.numeric(NA)}.
\item \code{prevalence}: a \code{\link[dplyr:reexports]{tibble}} listing the unique
epoch/periodicity found in \code{data}, along with its proportions.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{find_epoch()} returns the different epochs/periodicities/sample intervals
present in a time series, along with a best match epoch based on a predefined
threshold.
}
\details{
In the rare cases that a time series have periodicities with the same
prevalence above the the threshold, \code{find_epoch()} will return just one of
the values.
}
\examples{
index <- c(
as.POSIXct(seq(60, 5400, by = 60), origin = as.POSIXct("1970-01-01")),
as.POSIXct(seq(5430, 5490, by = 30), origin = as.POSIXct("1970-01-01")),
as.POSIXct(seq(5505, 5520, by = 15), origin = as.POSIXct("1970-01-01")),
as.POSIXct(seq(5530, 5540, by = 10), origin = as.POSIXct("1970-01-01")),
as.POSIXct(seq(5545, 5555, by = 5), origin = as.POSIXct("1970-01-01"))
)

data <- xts::xts(1:100, order.by = index)

find_epoch(data, 0.8)
}
\seealso{
Other utility functions: 
\code{\link{read_acttrust}()}
}
\concept{utility functions}
