% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_from_zenodo.R
\name{get_from_zenodo}
\alias{get_from_zenodo}
\title{Get data from a Zenodo record}
\usage{
get_from_zenodo(doi, dir = ".", file = NULL, parallel = FALSE, force = FALSE)
}
\arguments{
\item{doi}{A string indicating a Zenodo DOI (Digital Object Identifier)
starting with \code{"10.5281/zenodo."}. See the Examples section to learn more.}

\item{dir}{(optional) A string indicating a directory dir where the data
must be downloaded (default: \code{"."}).}

\item{file}{(optional) A \code{\link[base:character]{character}} vector with the
names of the files that must be downloaded. If \code{NULL}, the function will
download the entire record (default: \code{NULL}).}

\item{parallel}{(optional) A \code{\link[base:logical]{logical}} flag indicating if
the function must run a number of parallel processes, each downloading
another file. This is useful when multiple large files are present in the
Zenodo record, which otherwise would be downloaded sequentially. This
operation is limited by bandwidth and traffic limitations (default:
\code{FALSE}).}

\item{force}{(optional) A \code{\link[base:logical]{logical}} flag indicating if the
function should download the files even if they already exist
(default: \code{FALSE}).}
}
\value{
A \code{\link[=character]{character}} object with the paths for all the files
downloaded.
}
\description{
\code{get_from_zenodo()} allows you to easy download a Zenodo record or
a Zenodo file.

\strong{Note}: This function only works for Zenodo created DOIs (not when the DOI
is, for example, derived from Zookeys).
}
\details{
\subsection{Zenodo API}{

You can find more about the Zenodo API at:
\url{https://developers.zenodo.org}
}

\subsection{License}{

\code{get_from_zenodo()} code is based on the \code{download_zenodo()} function found
in the \href{https://github.com/inbo/inborutils}{\code{inborutils}} package of the
\href{http://www.inbo.be/en}{Research Institute for Nature and Forest (INBO)}.
\code{download_zenodo()} was created by Hans Van Calster (hans.vancalster@inbo.be)
and Floris Vanderhaeghe (floris.vanderhaeghe@inbo.be).

We give our thanks for the INBO institute and for all developers involved
in this piece of software.

Please note that this code comes with an \href{https://opensource.org/license/mit}{MIT License}. You can read the latter below.

\if{html}{\out{<div class="sourceCode">}}\preformatted{Copyright (c) 2016 Instituut voor Natuur en Bosonderzoek (INBO)

Permission is hereby granted, free of charge, to any person obtaining a copy
of this software and associated documentation files (the "Software"), to deal
in the Software without restriction, including without limitation the rights
to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
copies of the Software, and to permit persons to whom the Software is
furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in
all copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
SOFTWARE.
}\if{html}{\out{</div>}}
}
}
\examples{
\dontrun{
  library(readr)

  ## Downloading a single file from a Zenodo record

  dir <- tempdir()
  file <- "sleep-diary.txt"

  get_from_zenodo(
    doi = "10.5281/zenodo.4898822",
    dir = dir,
    file = file
  )

  read_lines(file.path(dir, file))

  ## Downloading all the files from a Zenodo record

  get_from_zenodo(
    doi = "10.5281/zenodo.4898822",
    dir = tempdir(),
    file = NULL
  )
}
}
\seealso{
Other API functions: 
\code{\link{get_sun_stats}()}
}
\concept{API functions}
