% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_data.R
\name{get_raw_data}
\alias{get_raw_data}
\title{Get paths to \code{actverse} raw datasets}
\usage{
get_raw_data(file = NULL)
}
\arguments{
\item{file}{(optional) A \code{\link[=character]{character}} vector specifying the raw
data file name(s) to retrieve. If \code{NULL}, returns all available raw data
file names (default: \code{NULL}).}
}
\value{
If \code{file} is \code{NULL}, a \code{\link[=character]{character}} object with all
file names available. Else, a string with the file name path.
}
\description{
\code{actverse} comes bundled with raw datasets for testing and learning
purposes. \code{get_raw_data()} makes it easy to access their paths.
}
\examples{
get_raw_data()

get_raw_data("acttrust.txt")
}
\seealso{
Other utility functions: 
\code{\link{aggregate_index}()},
\code{\link{find_epoch}()}
}
\concept{utility functions}
