% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sun_stats.R
\name{get_sun_stats}
\alias{get_sun_stats}
\title{Get sun related statistics from different APIs}
\usage{
get_sun_stats(
  latitude,
  longitude,
  date = Sys.Date(),
  tz = "UTC",
  method = "suncalc"
)
}
\arguments{
\item{latitude}{A number indicating the latitude of the desired location in
decimal degrees.}

\item{longitude}{A number indicating the longitude of the desired location in
decimal degrees.}

\item{date}{(optional) A \code{\link[base:as.Date]{Date}} value indicating the
moment in time (default: \code{Sys.Date()}).}

\item{tz}{(optional) A string indicating the time zone of the results. See
\code{\link[base:timezones]{timezones}} to learn more (default: \code{"UTC"}).}

\item{method}{(optional) A string indicating which method or API to use.
Valid values are: \code{"suncalc"} and \code{"sunrise-sunset.org"}. See the Details
section to learn more (default: \code{"suncalc"}).}
}
\value{
A \code{\link[=list]{list}} object with the following elements:
\itemize{
\item \code{date}: A \code{\link[base:as.Date]{Date}} vector with the same value of the
\code{date} parameter.
\item \code{latitude}: A number with the same value of the \code{latitude} parameter.
\item \code{longitude}: A number with the same value of the \code{longitude} parameter.
\item \code{tz}: A string with the same value of the \code{tz} parameter.
\item \code{sunrise_start}: An \code{\link[hms:hms]{hms}} value indicating the moment when
the top edge of the sun appears on the horizon.
\item \code{sunrise_end}: An \code{\link[hms:hms]{hms}} value indicating the moment when
bottom edge of the sun touches the horizon.
\item \code{golden_hour_end}: An \code{\link[hms:hms]{hms}} value indicating the moment when
the morning golden hour (soft light, best time for photography) ends.
\item \code{solar_noon}: An \code{\link[hms:hms]{hms}} value indicating the moment when sun
is in the highest position.
\item \code{golden_hour_start}: An \code{\link[hms:hms]{hms}} value indicating the moment
when the evening golden hour (soft light, best time for photography) starts.
\item \code{sunset_start}: An \code{\link[hms:hms]{hms}} value indicating the moment when the
bottom edge of the sun touches the horizon.
\item \code{sunset_end}: An \code{\link[hms:hms]{hms}} value indicating the moment when the
sun disappears below the horizon. This is also the moment when the evening
civil twilight starts.
\item \code{dusk}: An \code{\link[hms:hms]{hms}} value indicating the moment when the dusk
starts. This is also the moment when the evening nautical twilight starts.
\item \code{nautical_dusk}: An \code{\link[hms:hms]{hms}} value indicating the moment when
nautical dusk starts. This is also the moment when the evening astronomical
twilight starts.
\item \code{night_start}: An \code{\link[hms:hms]{hms}} value indicating the moment when
the night starts (dark enough for astronomical observations).
\item \code{nadir}: An \code{\link[hms:hms]{hms}} value indicating the darkest moment of the
night, i.e., when the sun is in the lowest position.
\item \code{night_end}: An \code{\link[hms:hms]{hms}} value indicating the moment when the
night ends. This is also the moment when the morning astronomical twilight
starts.
\item \code{nautical_dawn}:  An \code{\link[hms:hms]{hms}} value indicating the moment when
nautical dawn starts. This is also the moment when the morning nautical
twilight starts.
\item \code{dawn}: An \code{\link[hms:hms]{hms}} value indicating the moment when the dawn
starts. This is also the moment when the morning nautical twilight ends and
the morning civil twilight starts.
}
}
\description{
\code{get_sun_stats()} retrieves sun statistics using different models and
data sources.

At the moment, none of the \code{get_sun_stats()} methods uses real-world data.
All of them are based on models.

Each API have its peculiarities. We recommend checking the API documentation
for a better understanding of the mechanisms behind them (see the Details
section).
}
\details{
\subsection{\code{methods} argument}{

At the moment, \code{get_sun_stats()} can access the results of two APIs,
described below.
\itemize{
\item \code{"suncalc"}: Use the sun statistics provided by the
\href{https://github.com/datastorm-open/suncalc}{\code{suncalc}} package.
\item \code{"sunrise-sunset.org"}: Use the sun statistics provided by the
\href{https://sunrise-sunset.org/}{Sunrise-Sunset} API (requires an internet
connection). Click \href{https://sunrise-sunset.org/api}{here} to learn more.
}

The \code{"sunrise-sunset.org"} method tends to give a close, but usually lower,
result when compared with the \code{"suncalc"} method.

Please note that when using \code{method = "sunrise-sunset.org"} you need to show
attribution with a link to \url{https://sunrise-sunset.org}. Also note that
summer time adjustments are not included in the returned data when using this
method.
}

\subsection{Other statistics}{

The purpose of this function is to return basic statistics about the sun. If
you need other related statistics, we recommend checking the following
packages.
\itemize{
\item \href{https://github.com/rpradosiqueira/cptec/}{\code{cptec}}: An Interface to the
\href{https://www.cptec.inpe.br/}{CPTEC}/
\href{https://www.gov.br/inpe/}{INPE} API.
\item \href{https://docs.ropensci.org/nasapower/}{\code{nasapower}}: \href{https://power.larc.nasa.gov/}{NASA POWER} API Client.
\item \href{https://docs.ropensci.org/rnoaa/}{\code{rnoaa}}: \href{https://www.noaa.gov/}{NOAA}
Weather Data from R.
}
}
}
\examples{
latitude <- -23.5489
longitude <- -46.6388
date <- Sys.Date()
tz <- "America/Sao_Paulo"

get_sun_stats(
  latitude = latitude,
  longitude = longitude,
  date = date,
  tz = tz,
  method = "suncalc"
)

library(curl)

if (has_internet()) {
  get_sun_stats(
    latitude = latitude,
    longitude = longitude,
    date = date,
    tz = tz,
    method = "sunrise-sunset.org"
  )
}
}
\seealso{
Other API functions: 
\code{\link{get_from_zenodo}()}
}
\concept{API functions}
