% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npcra_is.R
\name{is}
\alias{is}
\title{Non-Parametric Function IS (Interdaily Stability)}
\usage{
is(x, timestamp, minutes_interval = 60)
}
\arguments{
\item{x}{Numeric vector with the activity data that will be used in the calculation.}

\item{timestamp}{POSIX vector that contains the date and time of each observation.}

\item{minutes_interval}{integer value representing the duration in minutes
of the time interval for grouping the data. By default, 60 minutes are
considered, this means that the activity will be averaged at hourly intervals.
The minimum value is 1 minute, where all points will be considered
(so there will be no average in intervals).}
}
\value{
A numeric value.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

Interdaily Stability identifies the synchronization with the 24h day-night cycle.
This fragmentation can have different results for the same data according
to the chosen time interval, that is, a minute by minute calculation brings a
more sensitive result than a check in hours, where small changes tend to have
less influence on interdaily stability
}
\examples{
\dontrun{
is()}
}
\references{
WITTING, W. et al. Alterations in the circadian rest-activity rhythm in aging
and Alzheimer's disease. Biological Psychiatry, v. 27, n. 6, p. 563-572,
Mar. 1990. doi: 10.1016/0006-3223(90)90523-5.

GONCALVES, Bruno S. B. et al. Nonparametric methods in actigraphy: an update.
Sleep Science, v. 7, n. 3, p. 158-164, 2014. doi: 10.1016/j.slsci.2014.09.013.
}
\seealso{
Other NPCRA functions: 
\code{\link{ism}()},
\code{\link{ivm}()},
\code{\link{iv}()},
\code{\link{l5_average_day}()},
\code{\link{l5_each_day}()},
\code{\link{npcra_l5_whole_period}()},
\code{\link{npcra_l5}()},
\code{\link{npcra_m10_each_day}()},
\code{\link{npcra_m10_whole_period}()},
\code{\link{npcra_m10}()},
\code{\link{ra}()}
}
\concept{NPCRA functions}
