% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_tidy}
\alias{is_tidy}
\title{Check if a data frame is tidy}
\usage{
is_tidy(data)
}
\arguments{
\item{data}{A data frame or tibble.}
}
\value{
A logic value indicating if the data frame is tidy.
}
\description{
To use this package is best that your data conforms to the proposed
data structure of the tidy_data function. You can still use the tools
available without tiding your data, but by doing that you may spend more
time using the functions, and may be more prone to error.
}
\examples{
\dontrun{
is_tidy(test_log)}
}
\seealso{
Other Utility functions: 
\code{\link{data_example}()},
\code{\link{model_data}()}
}
\concept{Utility functions}
