% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npcra_is.R
\name{ism}
\alias{ism}
\title{Non-Parametric Function ISm (Interdaily Stability mean)}
\usage{
ism(x, timestamp, minute_limit = 60)
}
\arguments{
\item{x}{Numeric vector with the activity data that will be used in the calculation.}

\item{timestamp}{POSIX vector that contains the date and time of each observation.}

\item{minute_limit}{integer value that corresponds to the last minute interval
to group the data. The default is 60, so 60 values of IS will be calculated to
take the average, with the first every minute and the last every 60 minutes.}
}
\value{
A numeric value.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

Interdaily Stability identifies the synchronization with the 24h day-night cycle.
This method calculates the average of ISs up to a minute limit. By default,
the limit is 60 minutes, so the 60 ISs will be calculated separately and the
results will be averaged to be returned.
}
\examples{
\dontrun{
ism()}
}
\references{
WITTING, W. et al. Alterations in the circadian rest-activity rhythm in aging
and Alzheimer's disease. Biological Psychiatry, v. 27, n. 6, p. 563-572,
Mar. 1990. doi: 10.1016/0006-3223(90)90523-5.

GONCALVES, Bruno S. B. et al. Nonparametric methods in actigraphy: an update.
Sleep Science, v. 7, n. 3, p. 158-164, 2014. doi: 10.1016/j.slsci.2014.09.013.
}
\seealso{
Other NPCRA functions: 
\code{\link{is}()},
\code{\link{ivm}()},
\code{\link{iv}()},
\code{\link{l5_average_day}()},
\code{\link{l5_each_day}()},
\code{\link{npcra_l5_whole_period}()},
\code{\link{npcra_l5}()},
\code{\link{npcra_m10_each_day}()},
\code{\link{npcra_m10_whole_period}()},
\code{\link{npcra_m10}()},
\code{\link{ra}()}
}
\concept{NPCRA functions}
