% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npcra_l5.R
\name{l5_average_day}
\alias{l5_average_day}
\title{Non-Parametric Function L5 (Least Active 5 Hours) for the avarage day}
\usage{
l5_average_day(x, timestamp)
}
\arguments{
\item{timestamp}{String with the name of the column that contains the date
and time of each observation (POSIX format).}

\item{data}{Dataframe that contains the date column and the column that
will be used to identify the least active 5-hour period.}

\item{col_activity}{String with the name of the column that will be used in the
calculation. The observations in this column must be in numeric format.}
}
\value{
A tibble with the value of L5 in the first position and the start
time (character format) of the 5-hour least active window in the period in the
second position.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

Calculates the least active window of 5 hours considering the average day of
the observations, that is, the window of the next 5 hours disregarding the
difference in days. The composition of all activities is considered
as a single day: the average day.
}
\examples{
\dontrun{
l5_average_day(test_log)}
}
\references{
WITTING, W. et al. Alterations in the circadian rest-activity rhythm in aging
and Alzheimer's disease. Biological Psychiatry, v. 27, n. 6, p. 563-572,
Mar. 1990. doi: 10.1016/0006-3223(90)90523-5.

GONCALVES, Bruno da Silva Brandao et al. A fresh look at the use of
nonparametric analysis in actimetry. Sleep Medicine Reviews, v. 20,
p. 84-91, Apr. 2015. doi: 10.1016/j.smrv.2014.06.002.
}
\seealso{
Other NPCRA functions: 
\code{\link{ism}()},
\code{\link{is}()},
\code{\link{ivm}()},
\code{\link{iv}()},
\code{\link{l5_each_day}()},
\code{\link{npcra_l5_whole_period}()},
\code{\link{npcra_l5}()},
\code{\link{npcra_m10_each_day}()},
\code{\link{npcra_m10_whole_period}()},
\code{\link{npcra_m10}()},
\code{\link{ra}()}
}
\concept{NPCRA functions}
