% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_tidy.R
\name{load_and_tidy}
\alias{load_and_tidy}
\title{Load and tidy an actimetry raw dataset}
\usage{
load_and_tidy(
  file = file.choose(),
  device = "acttrust1",
  tz = "America/Sao_Paulo"
)
}
\arguments{
\item{file}{A string with the csv file address for the raw dataset. If left
with no value, a dialog window will open allowing browsing and file
selection.}

\item{device}{A string with the name of the device used to generate the data.
See \code{\link[=tidy_data]{tidy_data()}} documentation to know the values available.}

\item{tz}{a character string that specifies which time zone to parse the
dates with. The string must be a time zone that is recognized by the
user's OS. For more information, see \code{?lubridate::ymd_hms}.}
}
\value{
A tibble.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

A wrapper for \link{load_data(}) and \code{\link[=tidy_data]{tidy_data()}} functions.

If you need to set specific parameters, please use the functions
mentioned above.
}
\examples{
\dontrun{
load_and_tidy(data_example("test_log.txt"))}
}
