% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_tidy.R
\name{load_data}
\alias{load_data}
\title{Load an actimetry raw dataset from a CSV file}
\usage{
load_data(
  file = file.choose(),
  device = "acttrust1",
  delim = ",",
  na = c("", " ", "NA"),
  skip = 0,
  skip_empty_rows = TRUE,
  trim_ws = TRUE
)
}
\arguments{
\item{file}{A string with the csv file address for the raw dataset. If left
with no value, a dialog window will open allowing browsing and file
selection.}

\item{device}{A string with the name of the device use to generate the file.
i.e. If this parameter is not \code{NULL}, all the other parameters, except
\code{file} will be overwritten. At the moment, the only value accepted is
\code{"acttrust1"}.}

\item{delim}{A string with the field separator in \code{file}.}

\item{na}{A character vector indicating values that must be interpreted
as \code{NA}.}

\item{skip}{A 0 or integer value with the number of rows to skip from file.}

\item{skip_empty_rows}{A logical value indicating if blank rows must be
ignored altogether. i.e. If this option is TRUE then blank rows will
not be represented at all. If it is \code{FALSE} then they will be represented
by \code{NA} values in all the columns.}

\item{trim_ws}{A logical value indicating if leading and trailing
whitespace must be trimmed from each field before parsing it.}
}
\value{
A tibble with all variables as character.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

This is a wrapper function to help simple data loading. It's based on
\code{\link[=read_delim]{read_delim()}} function found in the \code{readr} package (tidyverse). You don't need
to use this function if you already loaded your file.

If this function doesn't work for your file, we recommend using the \code{readr}
package to load it.
}
\examples{
\dontrun{
load_data(data_example("test_log.txt"))}
}
