% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npcra_is.R
\name{npcra_ism}
\alias{npcra_ism}
\title{Non-Parametric Function ISm (Interdaily Stability mean)}
\usage{
npcra_ism(tsbl, col, minute_limit = 60, summarize = FALSE)
}
\arguments{
\item{col}{A string indicating which column of \code{data} to use.}

\item{minute_limit}{integer value that corresponds to the last minute interval
to group the data. The default is 60, so 60 values of IS will be calculated to
take the average, with the first every minute and the last every 60 minutes.}

\item{summarize}{boolean value. If TRUE than returns just the ISm, if FALSE
returns a tibble with all values of IS between 1 and minute_limit.}

\item{data}{A \code{\link[tsibble:tsibble]{tsibble}} object.}
}
\value{
A numeric value.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Interdaily Stability identifies the synchronization with the 24h day-night
cycle.

This method calculates the average of IS's up to a minute limit. By default,
the limit is 60 minutes, so the 60 IS's will be calculated separately and the
results will be averaged to be returned.
}
\details{
Interdaily Stability (see \code{npcra_is()}) is a number between 0 and 1
calculated by dividing the square mean of the first derivative of the
data by the population variance (Witting, 1990).

From the IS stipulated by Witting, other estimates based on the IS were
derived, one being the mean IS (ISm). This method simply consists of
averaging IVs at different time intervals (Goncalves, 2014).
The function of this package considers a minute limit to calculate the
average of IS's.
As an example, the default is 60 minutes, so IVs will be calculated
with time intervals from 1 to 60 minutes and then average all these values.
}
\examples{
# Running for 1 week of random observations
first_date <- as.POSIXct('2015-01-01')
last_date <- as.POSIXct('2015-01-07')
shuffled_timestamp <- seq(first_date, last_date, by = "min")
timestamp <- sort(shuffled_timestamp)
x <- runif(8641, 0, 1000)
act <- dplyr::tibble(x, timestamp)
act <- tsibble::as_tsibble(act, index="timestamp")
npcra_ism(act, "x", minute_limit = 120, summarize=TRUE) # expects a numeric value
}
\references{
WITTING, W. et al. Alterations in the circadian rest-activity rhythm in aging
and Alzheimer's disease. Biological Psychiatry, v. 27, n. 6, p. 563-572,
Mar. 1990. doi: 10.1016/0006-3223(90)90523-5.

GONCALVES, Bruno S. B. et al. Nonparametric methods in actigraphy: an update.
Sleep Science, v. 7, n. 3, p. 158-164, 2014. doi: 10.1016/j.slsci.2014.09.013.
}
\seealso{
Other NPCRA functions: 
\code{\link{npcra_is}()},
\code{\link{npcra_ivm}()},
\code{\link{npcra_iv}()},
\code{\link{npcra_l5_each_day}()},
\code{\link{npcra_l5_whole_period}()},
\code{\link{npcra_l5}()},
\code{\link{npcra_m10_each_day}()},
\code{\link{npcra_m10_whole_period}()},
\code{\link{npcra_m10}()},
\code{\link{npcra_ra}()}
}
\concept{NPCRA functions}
