% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npcra.R
\name{npcra_iv}
\alias{npcra_iv}
\title{Non-Parametric Function IV (Intradaily  Variability)}
\usage{
npcra_iv(data, col_x = "pim")
}
\arguments{
\item{data}{Dataframe containing the column of numeric values that will be
used as X in the calculation.}

\item{col_x}{String with the name of the column that will be used in the
calculation.}
}
\value{
A numeric value.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

"IV was included to detect fragmentation of the rest-activity rhythms. High
IV may be an indicative of daytime napping and/or nighttime arousals" -
WITTING, W. (1990)
}
\examples{
\dontrun{
npcra_iv(test_log, "body_temperature")}

}
