% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npcra_iv.R
\name{npcra_ivm}
\alias{npcra_ivm}
\title{Non-Parametric Function IVm (Intradaily Variability mean)}
\usage{
npcra_ivm(data, col, minute_limit = 60, show_messages = TRUE, summarize = TRUE)
}
\arguments{
\item{data}{A \code{\link[tsibble:tsibble]{tsibble}} object.}

\item{col}{A string indicating which column of \code{data} to use.}

\item{minute_limit}{integer value that corresponds to the last
minute interval to group the data. The default is 60, so 60 values
of IV will be calculated to take the average, with the first
every minute and the last every 60 minutes.}

\item{show_messages}{if set to true it will return the IV for
every minute on the console as a message.}

\item{summarize}{if set to true, only the IVm value will be
output from the function, if set to false it will also return the
IV values for each minute in a tibble}
}
\value{
The numeric value of IVm if summarize is set as TRUE,
otherwise will be returned a tibble with the IVs until the minute limit with
their names (IVm, IV1, IV2, ...) in the first column and the values of IV
in the second column. The first column is called
iv_minute and the second as iv, with the IVm in the first row.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Intradaily Variability identifies the fragmentation of the
rest-activity rhythms.

This method calculates the average of IV's up to a minute limit. By default,
the limit is 60 minutes, so the 60 IV's will be calculated separately
and the results will be averaged to be returned.
}
\details{
Intradaily variability (see \code{npcra_iv()}) is a number that is
usually between 0 and 2 calculated by dividing the square mean of the
first derivative of the data by the population variance
(Witting et al., 1990).

Although real records typically have an IV value less than 2,
it is possible that this number exceeds 2. IVs greater than 2
are quite common in simulations of random values, since there is
no way to define a standard for the data and, therefore, will have
a high variability.

From the IV stipulated by Witting et al, other estimates based on the IV were
derived, one being the mean IV (IVm). This method simply consists of
averaging IVs at different time intervals (Goncalves et al., 2014).
The function of this package considers a minute limit to calculate the
average of IVs.
As an example, the default is 60 minutes, so IVs will be calculated
with time intervals from 1 to 60 minutes and then average all these values.
}
\examples{
#Running for 100 random observations
first_date <- as.POSIXct('2015-01-01')
last_date <- as.POSIXct('2015-01-11')
shuffled_timestamp <- sample(seq(first_date,
                      last_date, by = "sec"), 100)
timestamp <- sort(shuffled_timestamp)
x <- runif(n = 100, min = 0, max = 10000)

act <- dplyr::tibble(x, timestamp)
act <- tsibble::as_tsibble(act, index="timestamp")
npcra_ivm(act,
          col="x",
          minute_limit = 120,
          summarize = FALSE) #Expects a tibble 121 X 2
}
\references{
Witting, W., Kwa, I. H., Eikelenboom, P., Mirmiran, M., & Swaab, D. F.
(1990). Alterations in the circadian rest-activity rhythm in aging and
Alzheimer's disease. Biological Psychiatry, 27(6), 563-572.
\doi{https://doi.org/10.1016/0006-3223(90)90523-5}

Goncalves, B. S. B., Cavalcanti, P. R. A., Tavares, G. R., Campos,
T. F., & Araujo, J. F. (2014). Nonparametric methods in actigraphy: An
update. Sleep Science, 7(3), 158-164.
\doi{https://doi.org/10.1016/j.slsci.2014.09.013}
}
\seealso{
Other NPCRA functions:
\code{\link{npcra_iv}()}
}
\concept{NPCRA functions}
