% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npcra.R
\name{npcra_l5_whole_period}
\alias{npcra_l5_whole_period}
\title{Non-Parametric Function L5 (Least Active 5  Hours) for the entire period}
\usage{
npcra_l5_whole_period(
  data,
  col_name = "pim",
  timestamp = "timestamp",
  fast = TRUE
)
}
\arguments{
\item{data}{Dataframe that contains the date variable and the variable that
will be used to identify the least Active 5  Hours.}

\item{col_name}{String with the name of the column that will be used in the
calculation.}

\item{timestamp}{String with the name of the column that contains the
dataframe dates.}

\item{fast}{True: It makes the L5 quick calculation, prone to errors; False:
Scans completely and returns the correct L5 value, but more slowly}
}
\value{
a Dataframe with the value of L5 in the first position and the start date of the 5 least active window in the period in the second position.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

"L5 represents movement-activity during sleep plus nighttime arousals" -
WITTING, W. (1990) This calculation method considers the 5-hour window less
active for all data.
}
\examples{
\dontrun{
npcra_l5_whole_period(test_log, fast = FALSE)}

}
