% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npcra.R
\name{npcra_m10_each_day}
\alias{npcra_m10_each_day}
\title{Non-Parametric Function M10 (Most Active 10 Hours) for each day}
\usage{
npcra_m10_each_day(data, col_name = "pim", timestamp = "timestamp")
}
\arguments{
\item{data}{Dataframe that contains the date variable and the variable that
will be used to identify the 10 most active hours.}

\item{col_name}{String with the name of the column that will be used in the
calculation.}

\item{timestamp}{String with the name of the column that contains the
dataframe dates.}
}
\value{
a Dataframe with the value of M10 in the first position and the start
date of the 10 most active window by day in the second position.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

"M10 represents activity during the most active period of the day. This
measure may be influenced by daytime napping" - WITTING, W. (1990) This
calculation method considers the 10-hour window most active for each day.
}
\examples{
\dontrun{
npcra_m10_each_day(test_log, fast = FALSE)}
}
\seealso{
Other NPCRA functions: 
\code{\link{npcra_is}()},
\code{\link{npcra_iv}()},
\code{\link{npcra_l5_whole_period}()},
\code{\link{npcra_l5}()},
\code{\link{npcra_m10_whole_period}()},
\code{\link{npcra_m10}()},
\code{\link{npcra_test_args}()}
}
\concept{NPCRA functions}
