% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npcra.R
\name{npcra_test_args}
\alias{npcra_test_args}
\title{Testing the arguments of nonparametric functions}
\usage{
npcra_test_args(data, col_name, timestamp, method, fast)
}
\arguments{
\item{data}{Dataframe that contains the date variable and the variable that
will be used to identify the 10 most active hours.}

\item{col_name}{String with the name of the column that will be used in the
calculation.}

\item{timestamp}{String with the name of the column that contains the
dataframe dates.}

\item{method}{1 = whole period, 2 = average day, 3 = each day}

\item{fast}{True: It makes the M10 quick calculation, prone to errors; False:
Scans completely and returns the correct M10 value, but more slowly}
}
\value{
If no test stops, TRUE will be returned.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

Performs the verification of the types of expected values as arguments to the
npcra functions.
}
\examples{
\dontrun{
npcra_test_args(test_log, "pim", "timestamp", 1, TRUE)}


}
