% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npcra_ra.R
\name{ra}
\alias{ra}
\title{Non-Parametric Function RA (Relative Amplitude)}
\usage{
ra(x, timestamp, method = 1)
}
\arguments{
\item{x}{Numeric vector with the activity data that will be used in the calculation.}

\item{timestamp}{POSIX vector that contains the date and time of each observation.}

\item{method}{An integer that represents one of the three common methods for
calculating and analyzing RA:

1 = whole period,

2 = average day,

3 = each day.}
}
\value{
For methods 1 and 2 returns the numerical value that represents the
relative amplitude. For method equals 3 returns a tibble with the values of
RA in the first column and the date in the second column.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

Calculate the amplitude of the rest-activity rhythm. the result is based on
the difference between the 10 most active hours (M10) and the 5 least active hours (L5).
}
\examples{
\dontrun{
ra()}
}
\references{
WITTING, W. et al. Alterations in the circadian rest-activity rhythm in aging
and Alzheimer's disease. Biological Psychiatry, v. 27, n. 6, p. 563-572,
Mar. 1990. doi: 10.1016/0006-3223(90)90523-5.

GONCALVES, Bruno da Silva Brandao et al. A fresh look at the use of
nonparametric analysis in actimetry. Sleep Medicine Reviews, v. 20,
p. 84-91, Apr. 2015. doi: 10.1016/j.smrv.2014.06.002.
}
\seealso{
Other NPCRA functions: 
\code{\link{ism}()},
\code{\link{is}()},
\code{\link{ivm}()},
\code{\link{iv}()},
\code{\link{l5_average_day}()},
\code{\link{l5_each_day}()},
\code{\link{npcra_l5_whole_period}()},
\code{\link{npcra_l5}()},
\code{\link{npcra_m10_each_day}()},
\code{\link{npcra_m10_whole_period}()},
\code{\link{npcra_m10}()}
}
\concept{NPCRA functions}
