% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\name{raw_data}
\alias{raw_data}
\title{Get paths to \code{actverse} raw datasets}
\usage{
raw_data(file = NULL)
}
\arguments{
\item{file}{(optional) a \code{\link[=character]{character}} object indicating the raw
data file name(s). If \code{NULL}, all raw data file names will be printed
(default: \code{NULL}).}
}
\value{
\itemize{
\item If \code{file == NULL}, a \code{\link[=character]{character}} object with all file names
available.
\item Else, a string with the file name path.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

\code{actverse} comes bundled with raw datasets for testing and learning
purposes. \code{raw_data()} makes it easy to access their paths.
}
\examples{
## To list all raw data file names available

raw_data()

## To get the file path from a specific raw data

raw_data("acttrust.txt")
}
\seealso{
Other utility functions: 
\code{\link{aggregate_index}()},
\code{\link{find_epoch}()},
\code{\link{read_acttrust}()},
\code{\link{write_acttrust}()}
}
\concept{utility functions}
