% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sadeh_data.R
\name{sadeh_data}
\alias{sadeh_data}
\title{Classifies activity in stages of sleep and wakefulness using the Sadeh
algorithm}
\usage{
sadeh_data(data, year)
}
\arguments{
\item{data}{A data frame.}

\item{year}{Type of algorithm, published in the following year, 1 for the one published in 1989 and 2 for the one in 1994.}
}
\value{
The same data with two columns: STAGE and AWAKE.
}
\description{
This function uses a already loaded raw actimetry dataframe.
}
\details{
To use this package, the dataframe has to be tidy and organized, using the
tidy_data function You can still use the tools available without tiding your
data, but by doing that you may spend more time using the functions, and may
be more prone to error.
}
\note{
To do list:
\itemize{
\item Add new devices
\item Add parameters for unconformed data loading
}
}
\examples{
sadeh_data(data)
}
