% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep_statistic.R
\name{sleep_statistic}
\alias{sleep_statistic}
\title{Sleep statist}
\usage{
sleep_statistic(analyze, window_time = "00:30:00", pkg = "Y")
}
\arguments{
\item{analyze}{A matrix or array containing a vector with dates, vector with
times in the format 'hh:mm:ss', vector with states in the range of (1,2)
being: 1 - Sleep 2 - Resting, a vector with ZCM values, vector with PIM
values and vector with TAT values.}

\item{window_time}{A string, formated as \code{hh:mm:ss}, that sets the minimum
continuous value to set sleep. This will be use for calculate Sleep Onset
and Sleep Offset (O-O Interval).}

\item{pkg}{This parameter defines if the algorithm will calculate Mean and
Standard Deviation of the Circular time Series. This algorithm calculates
the variables "mean" and "Standard Deviation" using the Circular Package.
Note: If you don't need it,use as a value for the 'pkg' as 'N'. Valid
options: \code{"Y"} and \code{"N"}.}
}
\value{
A data frame containing the calculated variables:

Week Day, Date, Time in Bed (TIB), Sleep Period (O-O Interval), Sleep Minutes
during TIB (SMIN), True Sleep Minutes (TSMIN), Percent Sleep (PSLP), Sleep
Efficiency (SE), Mean Activity during TIB (AMEAN), Median Activity during TIB
(AMED), Activity Standard Deviation during TIB (ASD), Wake up After Sleep
Onset (WASO), Latency to Persistent Sleep (LPS), Wake Minutes during TIB
(WMIN).
}
\description{
Algorithm that calculates the main variables derived from the activity
log, based on the article 'Actigraphy-Based Assessment of Sleep Parameters'
by author Desta Fekedulegn.

Algorithm that calculates: Time in Bed (TIB), Sleep Period (O-O Interval),
Sleep Minutes during TIB (SMIN), True Sleep Minutes (TSMIN), Percent Sleep
(PSLP), Sleep Efficiency (SE), Mean Activity during TIB (AMEAN), Median
Activity during TIB (AMED), Activity Standard Deviation during TIB (ASD),
Wake up After Sleep Onset (WASO), Latency to Persistent Sleep (LPS) and Wake
Minutes during TIB (WMIN).

If you want to see more details about this paper, we recommend you to access
this link:
https://academic.oup.com/annweh/article/64/4/350/5735350?login=false
}
\details{
This algorithm uses the time circle of the data as a straight "line", aided
by the date and time.

There are two markers of this line in the algorithm, named "start" and
"finish". These markers help to identify the beginning and the end of the
sleep period respectively.

The "Start" marker has the function of assisting in marking the beginning of
sleep, using the "window_time" parameter. This marker will check if the next
few minutes (defined by the parameter) are marked as a sleep state. If so, it
marks the sleep start time, otherwise it advances to the next minute and
checks again.

The "Finish" marker has the function to help end sleep, as well as to change
days during the sleep period.

*Understand change of day in the sleep period as the sleep period that starts
on one day and ends on another.

Example: starts at 08:00 p.m. and ends at 06:00 a.m.
}
\concept{sleep statistics functions}
