% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sri.R
\name{sri}
\alias{sri}
\title{Compute Phillips et al.'s Sleep Regularity Index (SRI)}
\usage{
sri(
  data,
  state_col = "state",
  sleeping_states = c(1, 2),
  awake_states = 0,
  min_data = 0.75
)
}
\arguments{
\item{data}{A \code{\link[tsibble:tsibble]{tsibble}} object.}

\item{state_col}{(optional) A string indicating which column of \code{data} have
the states/categories data (default: \code{"state"}).}

\item{sleeping_states}{(optional) An
\code{\link[checkmate:checkIntegerish]{integerish}} vector indicating which state
values are considered as sleeping states (default: \code{c(1, 2)}).}

\item{awake_states}{(optional) An
\code{\link[checkmate:checkIntegerish]{integerish}} vector indicating which states
values are considered as awake states (default: \code{0}).}

\item{min_data}{(optional) A number indicating the minimum proportion of
non-missing values in the agreements required to compute the SRI for each
time point. The SRI will only be computed if the proportion of non-missing
values is greater than or equal to this threshold. This is useful to avoid
computing the SRI when there is insufficient data. Values below this
threshold will result in an SRI of \code{NA}. Note that the first agreement
value is always \code{NA} because there is no previous day for comparison, hence
it is not treated as missing data. (default: \code{0.75}).}
}
\value{
A \code{\link[tsibble:tsibble]{tsibble}} object with the following columns:
\itemize{
\item \code{time}: An \code{\link[hms:hms]{hms}} object representing the time of day.
\item \code{state}: A \code{\link[base:list]{list}} with a \code{\link[base:factor]{factor}} of the
sleep state at each time point ("Sleeping" or "Awake").
\item \code{previous_state}: A \code{\link[base:list]{list}} with a
\code{\link[base:factor]{factor}} of the sleep state at the same time on the
previous day.
\item \code{agreement}: A \code{\link[base:list]{list}} with a \code{\link[base:logical]{logical}}
vector indicating whether the state matches the previous day's state.
\item \code{sri}: A \code{\link[base:numeric]{numeric}} vector representing the Sleep
Regularity Index (SRI). See the Details section to learn more about how
the SRI is computed.
\item \code{valid_data}: A \code{\link[base:numeric]{numeric}} vector representing the
proportion of non-missing values in the \code{agreement} column, i.e., the
amount of information available to compute the SRI for each time point.
}
}
\description{
\code{sri()} computes the Sleep Regularity Index (SRI) as described by Phillips et
al. (\href{https://doi.org/10.1038/s41598-017-03171-4}{2017}) from a
\code{\link[tsibble:tsibble]{tsibble}} object, quantifying the consistency of
sleep/wake patterns across days.
}
\details{
The Sleep Regularity Index (SRI) is a measure of sleep regularity, which
quantifies the consistency of sleep patterns over time. It is calculated by
comparing the state of sleep at a given time of day with the state of sleep
at the same time on the previous day.

The index is computed as follows:
\enumerate{
\item For each time point, compare the sleep state (Sleeping/Awake) with the
state at the same time on the previous day.
\item Calculate the proportion of time points where the states match (i.e.,
agreement).
\item Scale this proportion to a range from -100 to 100: a score of 100
indicates perfectly regular sleep/wake patterns, while a score of 0
indicates completely random patterns.
}

Note 1: While SRI values below 0 are theoretically possible (e.g., in cases of
highly irregular patterns such as alternating 24-hour periods of sleep and
wake), such occurrences are extremely rare in practice.

Note 2: The first agreement value is always \code{NA} because there is no
previous day for comparison. This \code{NA} is not treated as missing data
because it simply reflects the absence of a reference point, and it is
excluded from the calculation of valid data.

See Phillips et al. (\href{https://doi.org/10.1038/s41598-017-03171-4}{2017}) to
learn more about the SRI and its applications in sleep research.
}
\examples{
library(curl)
library(dplyr)
library(ggplot2)
library(scales)

if (has_internet()) {
  file <- get_from_zenodo(
    doi = "10.5281/zenodo.4898822",
    dir = tempdir(),
    file = "processed.txt"
  )

  data <- read_acttrust(file, tz = "America/Sao_Paulo")
  sri_data <- data |> sri()

  sri_data |> print()

  sri_data |> pull(sri) |> summary() |> print()

  data |>
    actogram(
      col = "pim",
      days = -1,
      latitude = -23.55065,
      longitude = -46.63338,
      double_plot = TRUE
    )

  sri_data |>
    ggplot(ggplot2::aes(x = time, y = sri)) +
    geom_smooth(color = "#FF0000") +
    labs(
      x = "Time of Day (Hour)",
      y = "Sleep Regularity Index (SRI)"
    ) +
    scale_x_time(
      breaks = breaks_width("6 hours"),
      labels = label_time("\%-H") # Use "\%#H" for Windows
    ) +
    scale_y_continuous(limits = c(0, NA)) +
    actverse:::get_actverse_theme()
}
}
\references{
Phillips, A. J. K., Clerx, W. M., O’Brien, C. S., Sano, A., Barger, L. K.,
Picard, R. W., Lockley, S. W., Klerman, E. B., & Czeisler, C. A. (2017).
Irregular sleep/wake patterns are associated with poorer academic performance
and delayed circadian and sleep/wake timing. \emph{Scientific Reports}, \emph{7}(1),
3216. \doi{10.1038/s41598-017-03171-4}
}
\seealso{
Other sleep statistics functions: 
\code{\link{state_prop}()}
}
\concept{sleep statistics functions}
