% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_log}
\alias{test_log}
\title{An actimetric record}
\format{
A data frame with 10074 rows and 24 variables:

\describe{
\item{\strong{timestamp} (\code{POSIXct})}{
Log timestamp}

\item{\strong{ms} (\code{numeric})}{
Log millisecond}

\item{\strong{x_axis} (\code{numeric})}{
X axis signal}

\item{\strong{y_axis} (\code{numeric})}{
Y axis signal}

\item{\strong{z_axis} (\code{numeric})}{
Z axis signal}

\item{\strong{pim} (\code{numeric})}{
Wrist activity in Proportional Integral Mode (PIM)}

\item{\strong{pim_n} (\code{numeric})}{
Wrist activity in normalized PIM value per second (PIM divided by epoch
seconds)}

\item{\strong{tat} (\code{numeric})}{
Wrist activity in Time Above Threshold (TAT) mode}

\item{\strong{tat_n} (\code{numeric})}{
Wrist activity in normalized TAT value per second (TAT divided by epoch
seconds)}

\item{\strong{zcm} (\code{numeric})}{
Wrist activity in Zero Crossing Mode (ZCM)}

\item{\strong{zcm_n} (\code{numeric})}{
Wrist activity in normalized ZCM value per second (ZCM divided by epoch
seconds)}

\item{\strong{orientation} (\code{numeric})}{
Device orientation}

\item{\strong{body_temperature} (\code{numeric})}{
Body temperature in degrees Celsius (wrist skin temperature)}

\item{\strong{external_temperature} (\code{numeric})}{
External temperature in degrees Celsius}

\item{\strong{light} (\code{numeric})}{
Light intensity in lux}

\item{\strong{ambient_light} (\code{numeric})}
{Ambient light intensity in microwatts/square meter}

\item{\strong{red_light} (\code{numeric})}{
Red light intensity in microwatts/square centimeter}

\item{\strong{green_light} (\code{numeric})}{
Green light intensity in microwatts/square centimeter}

\item{\strong{blue_light} (\code{numeric})}{
Blue light intensity in microwatts/square centimeter}

\item{\strong{ir_light} (\code{numeric})}{
Infrared light intensity in microwatts/square centimeter}

\item{\strong{uva_light} (\code{numeric})}{
Ultraviolet A light intensity in microwatts/square centimeter}

\item{\strong{uvb_light} (\code{numeric})}{
Ultraviolet B light intensity in microwatts/square centimeter}

\item{\strong{event} (\code{logical})}{
A logical value indicating whether there was an event in the sampling
period}

\item{\strong{state} (\code{factor})}{
A factor value indicating the subject status: \code{"Awake"}, \code{"Resting"},
\code{"Sleeping"}, \code{"OffWrist"} (device removed) or editable states}
}
}
\source{
Prepared by Daniel Vartanian (package's author) (personal log).
}
\usage{
data(test_log)
}
\description{
An actimetric record, free for all kinds of use, personal made, and granted
by one of the authors of the package.

\strong{Device}: \href{https://www.condorinst.com.br/}{Condor Instruments} ActTrust 1
}
\references{
\strong{UNDER DEVELOPMENT}
}
\keyword{datasets}
