% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_tidy.R
\name{tidy_data}
\alias{tidy_data}
\title{Tidy an actimetry raw dataset}
\usage{
tidy_data(data, device = "acttrust1", tz = "America/Sao_Paulo")
}
\arguments{
\item{data}{A data frame or tibble.}

\item{device}{A string with the name of the device used to generate the data.
i.e. If this parameter is not \code{NULL}, all the other parameters, except
\code{tz} will be overwritten. At the moment, the only value accepted is
"acttrust1".}

\item{tz}{A character string that specifies which time zone to parse the
dates with. The string must be a time zone that is recognized by the
user's OS. For more information, see \code{?lubridate::ymd_hms}.}
}
\value{
A tibble.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

This function tidy a already loaded raw actimetry dataset. If your data
ain't loaded yet, please use the \code{\link[=load_data]{load_data()}} function or load it by
yourself.

To use this package is best that your data conforms to the proposed
data structure of this function. You can still use the tools available
without tiding your data, but by doing that you may spend more time using
the functions, and may be more prone to error.
}
\note{
To do list:
\itemize{
\item Add new devices
\item Add parameters for unconformed data loading
}
}
\examples{
\dontrun{
tidy_data(load_data(data_example("test_log.txt")))}
}
