% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webster.R
\name{webster}
\alias{webster}
\title{Classifies activity using the Webster algorithm}
\usage{
webster(data)
}
\arguments{
\item{data}{A \code{data.frame} object.}
}
\value{
The same data with two columns: STAGE and AWAKE.
}
\description{
This function uses a already loaded raw actimetry dataframe.
}
\examples{
\dontrun{
data <- readr::read_csv2("./inst/extdata/test_log.txt", skip = 25)
data <- webster(data)
data$timestamp = lubridate::dmy_hms(paste(data$DATE, data$TIME))
ggplot2::qplot(timestamp, AWAKE, data = data, geom = "area",
               xlab = "Timestamp", ylab = "State")
}
}
