% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webster_data.R
\name{webster_data}
\alias{webster_data}
\title{Classifies activity in stages of sleep and wakefulness using the Webster algorithm}
\usage{
webster_data(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
The same data with two columns: STAGE and AWAKE.
}
\description{
This function uses a already loaded raw actimetry dataframe.
}
\details{
To use this package, the dataframe has to be tidy and organized, using the tidy_data function
You can still use the tools available without tiding your data,
but by doing that you may spend more time using the functions, and may be more prone to error.
}
\note{
To do list:
\itemize{
\item Add new devices
\item Add parameters for unconformed data loading
}
}
\examples{
webster_data(data)
}
