% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_acttrust.R
\name{write_acttrust}
\alias{write_acttrust}
\title{Adapt and write a time series to a readable ActTrust file}
\usage{
write_acttrust(data, file, delim = ";", zero_offwrist = FALSE, header = NULL)
}
\arguments{
\item{data}{A \code{\link[tsibble:tsibble]{tsibble}} object with a \code{POSIXct} or
\code{POSIXt} index. The ActTrust software only deals with this type of index.}

\item{file}{A string with a file path to write to.}

\item{delim}{(optional) a string indicating the delimiter that must be used
to separate values. Valid delimiters: \code{";"} and \code{"\\t"} (default: \code{";"}).}

\item{zero_offwrist}{(optional) a \code{\link[=logical]{logical}} value indicating if
all values classified as offwrist (device removal) must be assigned with
\code{0} (default: \code{FALSE}).}

\item{header}{(optional) a string indicating the file path to a file (usually
the raw data file) with the ActTrust header. This is not mandatory,
the ActTrust software can read files without it (default: \code{NULL}).}
}
\value{
An invisible \code{NULL}. This function don't aim to return values.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{write_acttrust()} allows you to adapt and write a
\code{\link[tsibble:tsibble]{tsibble}} object to a readable ActTrust file.

Note that your data must conform to a predefined structure. See
\code{\link[=acttrust]{?acttrust}} to learn more.

ActTrust is a trademark of
\href{https://www.condorinst.com.br/}{Condor Instruments Ltda}.
}
\details{
\code{write_acttrust()} will transform any \code{NA} value to \code{0}. This is because
the ActTrust software only interpret numeric values.
}
\examples{
\dontrun {
file <- tempfile()
header <- raw_data("acttrust.txt")

write_acttrust(acttrust, file, header = header)

readLines(file, n = 30)
}
}
\seealso{
Other utility functions: 
\code{\link{aggregate_index}()},
\code{\link{find_epoch}()},
\code{\link{raw_data}()},
\code{\link{read_acttrust}()}
}
\concept{utility functions}
