% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acttrust.R
\docType{data}
\name{acttrust}
\alias{acttrust}
\title{An ActTrust actigraphy record}
\format{
A \code{\link[tsibble:tsibble]{tsibble}} with 17 columns and 1,441 rows:

\describe{
\item{timestamp}{
Date and time of the recorded events.
\cr \cr
R class: \code{\link[=as.POSIXct]{POSIXct}}.
}

\item{pim}{
Wrist activity in Proportional Integral Mode (PIM).
\cr \cr
R class: \code{\link[=numeric]{numeric}}.
}

\item{tat}{
Wrist activity in Time Above Threshold (TAT) mode.
\cr \cr
R class: \code{\link[=numeric]{numeric}}.
}

\item{zcm}{
Wrist activity in Zero Crossing Mode (ZCM).
\cr \cr
R class: \code{\link[=numeric]{numeric}}.
}

\item{orientation}{
Device orientation.
\cr \cr
R class: \code{\link[=numeric]{numeric}}.
}

\item{wrist_temperature}{
Wrist temperature in degrees \code{Celsius}.
\cr \cr
R class: \code{\link[=numeric]{numeric}}.
}

\item{external_temperature}{
External temperature in degrees \code{Celsius} (unreliable data).
\cr \cr
R class: \code{\link[=numeric]{numeric}}.
}

\item{light}{
Light intensity in \code{lux}.
\cr \cr
R class: \code{\link[=numeric]{numeric}}.
}

\item{ambient_light}{
Ambient light intensity in \verb{microwatts/square meter}.
\cr \cr
R class: \code{\link[=numeric]{numeric}}.
}

\item{red_light}{
Red light intensity in \verb{microwatts/square centimeter}.
\cr \cr
R class: \code{\link[=numeric]{numeric}}.
}

\item{green_light}{
Green light intensity in \verb{microwatts/square centimeter}.
\cr \cr
R class: \code{\link[=numeric]{numeric}}.
}

\item{blue_light}{
Blue light intensity in \verb{microwatts/square centimeter}.
\cr \cr
R class: \code{\link[=numeric]{numeric}}.
}

\item{ir_light}{
Infrared light intensity in \verb{microwatts/square centimeter}.
\cr \cr
R class: \code{\link[=numeric]{numeric}}.
}

\item{uva_light}{
Ultraviolet A light intensity in \verb{microwatts/square centimeter}.
\cr \cr
R class: \code{\link[=numeric]{numeric}}.
}

\item{uvb_light}{
Ultraviolet B light intensity in \verb{microwatts/square centimeter}.
\cr \cr
R class: \code{\link[=numeric]{numeric}}.
}

\item{event}{
A binary number indicating the moments when the subject pressed the event
button: \code{0} when the button is not pressed and \code{1} when the button is
pressed. This is commonly used to indicate when the subject goes to sleep
or wakes up.
\cr \cr
R class: \code{\link[=numeric]{numeric}}.
}

\item{state}{
An integer number indicating the subject state: \code{0} for awake, \code{1} for
sleeping, \code{2} for resting, \code{4} for offwrist (device removal), \code{6} for
editable state 1, \code{7} for editable state 2, and \code{8} for editable state 3.
These states can be assigned by a specialist or by a validated inference
algorithm.
\cr \cr
R class: \code{\link[=numeric]{numeric}}.
}
}
}
\source{
Created by Daniel Vartanian (package author).
}
\usage{
acttrust
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

A small sample (1 day length) of an
\href{https://condorinst.com/en/acttrust-actigraph/}{ActTrust}
actigraphy record.

\strong{Note}: We're always looking for actigraphy data samples. If you have a
personal actigraphy record made with a device that is not listed in the
\code{actverse} package and is ok with sharing it, please contact the package
authors.

ActTrust is a trademark of
\href{https://condorinst.com/en/}{Condor Instruments Ltda}.
}
\details{
This dataset was created by one of the \code{actverse} package authors and is free
for all kinds of use. You can download the full dataset at
\url{https://zenodo.org/record/4898822}.
\subsection{Data wrangling}{

This data was tidied and validated with the default settings of
\code{\link[=read_acttrust]{read_acttrust()}}.

The process of \emph{tiding} a dataset is understood as transforming it in input
data, like described in Loo and Jonge (2018). It's a very similar process of
tiding data described in the workflow proposed by Wickham and Grolemund
(n.d.).

The process of \emph{validating} a dataset is understood as detecting invalid
data, by checking whether data satisfies certain assumptions from domain
knowledge, to then,  removing or, if possible, fixing them. This process can
be considered as part of the process of transforming data, described in the
workflow proposed by Wickham and Grolemund (n.d.).

To learn more about the concept of tidy data, see Wickham (2014) and Wickham
and Grolemund (n.d.). You can find more about data validation and error
location in Loo and Jonge (2018).
}

\subsection{Variable naming}{

The \code{actverse} package stick to the \href{https://tidyverse.tidyverse.org/articles/manifesto.html}{tidyverse principles}. Our
code and naming schemes follow the guidelines of the \href{https://style.tidyverse.org/}{tidyverse style guide}.
}

\subsection{Variable classes}{

R has a \href{https://cran.r-project.org/view=TimeSeries}{vast number} of time
series standards. We choose to work with \code{\link[tsibble:tsibble]{tsibble}}
because we think is the best standard that follows the \href{https://tidyverse.tidyverse.org/articles/manifesto.html}{tidyverse principles}.

To make sure that the resulting data from \code{actverse} could be easily
transformed to other standards, we decided to maintain all variable classes
as \code{\link[=numeric]{numeric}}, except for the index (class
\code{\link[=as.POSIXct]{POSIXct}}).

If you're looking for a good way to easily transform your data to other time
series standards, check the \href{https://www.tsbox.help/}{\code{tsbox}} package.
}

\subsection{Raw data}{

The raw data for this dataset can be found with
\code{\link[=raw_data]{raw_data("acttrust.txt")}}.
}
}
\section{Metadata}{

\subsection{Subject}{
\itemize{
\item Age: \code{35}.
\item Gender: Male.
\item Height: \code{180} cm.
\item Weight: \code{85} kg.
\item BMI: \code{~26.235} (overweight).
}
}

\subsection{Record}{
\itemize{
\item Body part: Left wrist (non-dominant wrist).
\item Start: \verb{2021-04-24 04:14:25}.
\item End: \verb{2021-05-31 03:39:06}.
\item Days: \code{37}
\item Epoch (sample interval): \code{60} seconds.
\item Mode: PIM, TAT, and ZCM.
\item TAT Threshold: \code{1024}.
\item Location: Sao Paulo, SP, Brazil.
\item UTC: \code{-3}.
\item DST: \code{FALSE}.
\item Light phase: \code{06:23:59} - \code{17:44:58}.
\item Dark phase: \code{17:44:59} - \code{06:23:58}.
}
}

\subsection{Actigraph}{
\itemize{
\item Brand: \href{https://condorinst.com/en/}{Condor Instruments}.
\item Model: \href{https://condorinst.com/en/acttrust-actigraph/}{ActTrust 1}.
\item Hardware version: \code{3.1}.
\item Firmware version: \code{3.7}.
\item Manual: \verb{1.0.13} (PT-BR).
\item Software: ActStudio Alpha 1.0.17 (Win | Mac).
}
}

\subsection{Scoring settings}{
\itemize{
\item Scoring algorithm: Condor Instruments.
\item The sleep diary data was not used for scoring.
}
}

\subsection{Scoring review}{
\itemize{
\item Offwrist periods were indicated from sudden drops in wrist temperature.
\item Micro sleep states (isolated states lasting less than \verb{5 minutes}) were
removed.
\item Minor adjustments were made on the sleep onset and sleep offset, taking
into consideration the event button and sleep diary records.
}
}

\subsection{Sleep statistics settings}{
\itemize{
\item Multiple main sleep periods: \code{TRUE}.
\item Main sleep period time of day: \verb{Day or Night}.
\item Minimum duration for main sleep periods: \code{01:00:00}.
\item Awake period limit: \code{02:00:00}.
}
}
}

\references{
Condor Instruments. (2019). \emph{ActTrust - user manual}. Condor Instruments.
Model AT0503. ActStudio version 1.0.13.

van der Loo, M., & de Jonge, E. (2018). \emph{Statistical data cleaning with
applications in R}. John Wiley & Sons. \doi{10.1002/9781118897126}.

Wickham, H. (2014). Tidy Data. \emph{Journal of Statistical Software}, \emph{59}(10),
1-23. \doi{10.18637/jss.v059.i10}.

Wickham, H., & Grolemund, G. (n.d.). \emph{R for data science}. (n.p.).
\url{https://r4ds.had.co.nz}
}
\concept{datasets}
\keyword{datasets}
