% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actverse-package.R
\docType{package}
\name{actverse-package}
\alias{actverse}
\alias{actverse-package}
\title{actverse: The R Package for Actigraphic Data}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A set of tools for processing, analyzing and visualizing 
    actigraphy data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gipsousp.github.io/actstudio/}
  \item \url{https://github.com/gipsousp/actstudio/}
  \item Report bugs at \url{https://github.com/gipsousp/actstudio/issues}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID})

Authors:
\itemize{
  \item Bruno Furlan \email{brfrlan@gmail.com} (\href{https://orcid.org/0000-0003-1654-0899}{ORCID})
  \item Vinicius Matias \email{viniciusmatias@usp.br} (\href{https://orcid.org/0000-0002-2047-5420}{ORCID})
  \item Cassio Serrano \email{cassio.serrano@usp.br} (\href{https://orcid.org/0000-0001-7149-1829}{ORCID})
  \item Ana Amelia Benedito-Silva \email{aamelia@usp.br} (\href{https://orcid.org/0000-0003-4976-2623}{ORCID}) [expert, scientific advisor]
  \item Mario Pedrazzoli \email{pedrazzo@usp.br} (\href{https://orcid.org/0000-0002-5257-591X}{ORCID}) [expert, scientific advisor]
}

Other contributors:
\itemize{
  \item GIPSO [copyright holder, funder]
  \item USP [funder]
}

}
\keyword{internal}
