% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actverse-package.R
\docType{package}
\name{actverse-package}
\alias{actverse}
\alias{actverse-package}
\title{actverse: Actigraphy Data Analysis Tools}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A complete and consistent toolkit to process, analyze and visualize actigraphy data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gipso.github.io/actverse/}
  \item \url{https://github.com/gipso/actverse/}
  \item Report bugs at \url{https://github.com/gipso/actverse/issues}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Vinicius Alves Matias \email{viniciusmatias@usp.br} (\href{https://orcid.org/0000-0002-2047-5420}{ORCID})
  \item Cassio Almeida Mattos Serrano \email{cassio.serrano@usp.br} (\href{https://orcid.org/0000-0001-7149-1829}{ORCID})
  \item Ana Amelia Benedito-Silva \email{aamelia@usp.br} (\href{https://orcid.org/0000-0003-4976-2623}{ORCID}) [scientific advisor]
  \item Mario Pedrazzoli \email{pedrazzo@usp.br} (\href{https://orcid.org/0000-0002-5257-591X}{ORCID}) [scientific advisor]
}

Other contributors:
\itemize{
  \item Interdisciplinary Sleep Research Group (GIPSO) [funder]
  \item University of Sao Paulo (USP) [funder]
  \item Bruno Ramalho Furlan \email{brfrlan@gmail.com} (\href{https://orcid.org/0000-0003-1654-0899}{ORCID}) [contributor]
}

}
\keyword{internal}
