% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_example}
\alias{data_example}
\title{Get path to data example}
\usage{
data_example(path = NULL)
}
\arguments{
\item{path}{A string with the file name. If \code{NULL}, the example
files will be listed.}
}
\value{
If path is equal to \code{NULL}, returns a character vector with all
the example file names. Else, returns the full path where the file is
located.
}
\description{
Actverse comes bundled with a number of sample files in its \code{inst/extdata}
directory. This function make them easy to access.
}
\examples{
\dontrun{
data_example()
data_example("test_log.txt")}
}
