% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epoch_change.R
\name{epoch}
\alias{epoch}
\title{Return the epoch length of an actimetry dataset}
\usage{
epoch(data, timestamp = "timestamp")
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble} object.}

\item{timestamp}{A string indicating the name of the timestamp
variable. The timestamp variable must be class \code{POSIXct}.}
}
\value{
A numeric value representing the epoch length in seconds.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
epoch(test_log)
}
}
