% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_epoch.R
\name{find_epoch}
\alias{find_epoch}
\title{Find the epochs/periodicities in a \code{tsibble}}
\usage{
find_epoch(data, threshold = 0.9)
}
\arguments{
\item{data}{A \link[tsibble:tsibble]{tsibble} object with a \code{\link[=as.Date]{Date}}
or \code{\link[=as.POSIXct]{POSIXt}} vector as index.}

\item{threshold}{(optional) a number, from \code{0} to \code{1}, indicating the minimum proportion
that an epoch must have to be considered valid. \code{threshold = 1} means that
the regularity of the time series must be strict (i.e., have just 1
periodicity) (default: \code{0.9}).}
}
\value{
A \code{\link[=list]{list}} object with the following elements:
\itemize{
\item \code{best_match}: A number indicating the epoch/periodicity above the
\code{threshold} with greater prevalence in seconds. If none is find, \code{best_match}
value will be equal as \code{as.numeric(NA)}.
\item \code{prevalence}: a \code{\link[tibble:tibble]{tibble}} listing the unique
epochs/periodicities found in \code{data} along with its proportions.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{find_epoch()} returns the different epochs/periodicities present in a
\code{\link[tsibble:tsibble]{tsibble}} object along with a best match epoch based
on a predefined threshold.
}
\details{
In the rare cases that a \code{\link[tsibble:tsibble]{tsibble}} object have
periodicities with the same prevalence above the threshold, \code{best_match} will
return just one of those values.
}
\examples{
data <- dplyr::tibble(
    index = c(
        as.POSIXct(seq(60, 5400, by = 60), origin = lubridate::origin),
        as.POSIXct(seq(5430, 5490, by = 30), origin = lubridate::origin),
        as.POSIXct(seq(5505, 5520, by = 15), origin = lubridate::origin),
        as.POSIXct(seq(5530, 5540, by = 10), origin = lubridate::origin),
        as.POSIXct(seq(5545, 5555, by = 5), origin = lubridate::origin)
        ),
    x = seq_along(timestamp)
    )
data <- tsibble::tsibble(data, index = index)

find_epoch(data, 0.8)
}
\seealso{
Other utility functions: 
\code{\link{aggregate_index}()},
\code{\link{raw_data}()},
\code{\link{read_acttrust}()},
\code{\link{write_acttrust}()}
}
\concept{utility functions}
