% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{model_data}
\alias{model_data}
\title{Return a tidy data frame}
\usage{
model_data()
}
\value{
A tibble.
}
\description{
This function return a data frame conformed to the proposed data structure
of the \code{\link[=tidy_data]{tidy_data()}} function. It can be used for testing and structuring
other routines.
}
