% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npcra.R
\name{npcra_is}
\alias{npcra_is}
\title{Non-Parametric Function IS (Interdaily Stability )}
\usage{
npcra_is(data, col_x = "pim", col_date = "timestamp")
}
\arguments{
\item{data}{Dataframe containing the column of numeric values that will be
used as X in the calculation.}

\item{col_x}{String with the name of the column that will be used in the
calculation.}

\item{col_date}{String with the name of the column with date values.}
}
\value{
A numeric value.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

"Alterations in IS may indicate a  loose coupling between the rest-activity
rhythm and its supposedly stable "Zeitgebers," as IS decreases with higher
day-to-day variation (i.e. ,  a  looser coupling) of the activity patterns" -
WITTING, W. (1990)
}
\examples{
\dontrun{
npcra_is(test_log, "pim", "timestamp")}
}
\seealso{
Other NPCRA functions: 
\code{\link{npcra_iv}()},
\code{\link{npcra_l5_whole_period}()},
\code{\link{npcra_l5}()},
\code{\link{npcra_m10_each_day}()},
\code{\link{npcra_m10_whole_period}()},
\code{\link{npcra_m10}()},
\code{\link{npcra_test_args}()}
}
\concept{NPCRA functions}
