% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npcra_iv.R
\name{npcra_iv}
\alias{npcra_iv}
\title{Non-Parametric Function IV (Intradaily Variability)}
\usage{
npcra_iv(x, timestamp, minutes_interval = 60)
}
\arguments{
\item{x}{Numeric vector with the activity data that will be used in the
calculation.}

\item{timestamp}{POSIX vector that contains the date and time of each
observation.}

\item{minutes_interval}{integer value representing the duration in minutes
of the time interval for grouping the data. By default, 60 minutes are
considered, this means that the activity will be averaged at hourly intervals.
The minimum value is 1 minute, where all points will be considered
(so there will be no average in intervals).}
}
\value{
A numeric value.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Calculates the Intradaily Variability, which identifies the
fragmentation of the rest-activity rhythms.

This fragmentation can have different results for the same data according
to the chosen time interval, that is, a minute by minute calculation brings
a more sensitive result than a check in hours, where small changes tend
to have less influence on intradaily variability.
}
\details{
Intradaily variability is a number that is usually between 0 and
2 calculated by dividing the square mean of the first derivative
of the data by the population variance (Witting, 1990).

Although real records typically have an IV value less than 2,
it is possible that this number exceeds 2. IVs greater than 2
are quite common in simulations of random values, since there is
no way to define a standard for the data and, therefore, will have
a high variability.

If X represents the activity or other numeric variable passed as parameter
\code{x}, the IV follows the formula below:

\deqn{IV = SQM_FD / VAR(X)}

Where SQM_FD is the square mean of the first derivative of the data,
calculated as below:

\deqn{SQM_FD = \sum_{i=2}^{N} (X_i - X_{i-1})^2 / (N-1)}

And the population variance VAR(X) is:

\deqn{VAR(X) = \sum_{i=1}^{N} (\bar{X} - X_i)^2}

Where:

\eqn{X_i} is each data point or, more likely, each activity average
for the time interval passed by the parameter \code{minutes_interval}

N is the amount of data or the number of intervals that can fit in
the data set

\eqn{\bar{X}} is the average of all data or the mean of all average
activities by time interval

Usually the activity data used in the calculation are hourly averages of
the activity, thus avoiding activity fluctuations in the same period of
time. The calculation for IV by hourly averages is also called IV60
(Goncalves et al., 2014), this being the standard minute interval for
the method (60 minutes). It is still possible to vary this interval of
minutes, which can generate totally different results for the IV that
can be analyzed to identify some pattern of the data.

Higher values of IV represent a greater fragmentation of the rest - activity
rhythm, this is because the calculation has a certain sensitivity to
immediate changes between time intervals, such as naps during the day and
nighttime awakenings.
}
\examples{
#Using the test_log data from the package
npcra_iv(test_log$pim, test_log$timestamp)
# [1] 0.7290303

#Running for 100 random observations
first_date <- as.POSIXct('2015-01-01')
last_date <- as.POSIXct('2015-01-11')
shuffled_timestamp <- sample(seq(first_date,
                      last_date, by = "sec"), 100)
timestamp <- sort(shuffled_timestamp)
x <- runif(100, 0, 10000)
npcra_iv(x, timestamp, minutes_interval = 120) #expects a numeric  value
}
\references{
WITTING, W. et al. Alterations in the circadian rest-activity rhythm in aging
and Alzheimer's disease. Biological Psychiatry, v. 27, n. 6, p. 563-572,
Mar. 1990. doi: 10.1016/0006-3223(90)90523-5.

GONÇALVES, Bruno S. B. et al. Nonparametric methods in actigraphy: an update.
Sleep Science, v. 7, n. 3, p. 158-164, 2014. doi: 10.1016/j.slsci.2014.09.013.
}
\seealso{
Other NPCRA functions:
\code{\link{npcra_ivm}()}
}
\concept{NPCRA functions}
