% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npcra_l5.R
\name{npcra_l5_each_day}
\alias{npcra_l5_each_day}
\title{Non-Parametric Function M10 (Least Active 5 Hours) for each day}
\usage{
npcra_l5_each_day(x, timestamp)
}
\arguments{
\item{x}{Numeric vector with the activity data that will be used in the
calculation.}

\item{timestamp}{POSIXct/POSIXlt vector that contains the date and time of
each observation.}
}
\value{
A tibble of two columns, with the values of L5 in the first column
and the start dates of the 5 least active period for each day in the second
position (so, each row is a different day).
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

Calculates and finds the least active period of 5 hours for each different
day.
}
\details{
The L5 is the smallest average activity over a 5-hour period.
The first uses of the l5 consisted of calculating the 5 least active hours
based on the hourly means (WITTING et al., 1990). As it does not have an
explicit definition, it is expected that the values would be calculated
under the complete period (see \code{npcra_l5_whole_period()}) received
and that is why that is why many authors perform the analysis
by two other approaches (GONCALVES et al., 2015): calculation for each
day and for the 24h mean profile (see \code{npcra_l5_mean_profile()})

l5 for each day calculates the least active 5 hour period for each
different date (day, month and year). Thus, each date of the data set
delivered must have a value of l5 as long as there is a complete period
of 5 hours for that date. The results can quantify the intensity of nocturnal
activity (the higher the result, the more intense) and variations by date.

The function receives two vectors, one containing the registered activity
(numerical values) and another vector containing the date of each occurrence
(POSIXct / POSIXlt), so it is possible to map the activity and dates by the
indexes of the vectors.

Since the vectors are organized according to the date of observation in
ascending order (see Note), the function will identify the date of the last
full 10-hour period in the record.

With this information the verification of all valid activity windows for each
date starts.

First all different dates are identified and what activity data belongs to
them. In sequence, the averages of the 5-hour windows of the first day
are calculated, that is, the averages of all windows that end at 23:59:59
or earlier. This process is repeated for the next day and ends when the
last valid date arrives.

At the end of the calculation, the lowest average for its respective day
is captured
}
\note{
As normally expected, the date array must be ordered in ascending order.
This allows the verification of the longest periods to follow a method of
asymptotic complexity O(2n) = O(n), where n is the amount of data
(activity / date), so for 10,000 data, a little less than 20,000 interactions
will be performed, reducing the time to obtain the result for seconds.
}
\examples{
#Using the test_log data from the package
npcra_l5_each_day(test_log$pim, test_log$timestamp)

#Running for 1000 random observations
first_date <- as.POSIXct('2015-01-01')
last_date <- as.POSIXct('2015-01-11')
shuffled_timestamp <- sample(seq(first_date,
                                 last_date, by = "min"), 1000)
timestamp <- sort(shuffled_timestamp)
x <- runif(1000, 0, 10000)
npcra_l5_each_day(x, timestamp) #returns a tibble 10X2

}
\references{
WITTING, W. et al. Alterations in the circadian rest-activity rhythm in aging
and Alzheimer's disease. Biological Psychiatry, v. 27, n. 6, p. 563-572,
Mar. 1990. doi: 10.1016/0006-3223(90)90523-5.

GONCALVES, Bruno da Silva Brandao et al. A fresh look at the use of
nonparametric analysis in actimetry. Sleep Medicine Reviews, v. 20,
p. 84-91, Apr. 2015. doi: 10.1016/j.smrv.2014.06.002.
}
\seealso{
Other NPCRA functions: 
\code{\link{ism}()},
\code{\link{is}()},
\code{\link{ivm}()},
\code{\link{iv}()},
\code{\link{npcra_l5_whole_period}()},
\code{\link{npcra_l5}()},
\code{\link{npcra_m10_each_day}()},
\code{\link{npcra_m10_whole_period}()},
\code{\link{npcra_m10}()},
\code{\link{npcra_ra}()}
}
\concept{NPCRA functions}
