% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npcra_m10.R
\name{npcra_m10}
\alias{npcra_m10}
\title{Non-Parametric Function M10 (Most Active 10 Hours)}
\usage{
npcra_m10(x, timestamp, method = 1)
}
\arguments{
\item{x}{Numeric vector with the activity data that will be used in the
calculation.}

\item{timestamp}{POSIX vector that contains the date and time of each
observation.}

\item{method}{An integer that represents one of the three common methods for
calculating and analyzing M10:

1 = whole period,

2 = average day,

3 = each day.

If you prefer you can also directly call the functions
\code{npcra_m10_whole_period()}, \code{npcra_m10_average_day()}
or \code{npcra_m10_each_day()} and see the differences in the calculation
method.}
}
\value{
A tibble with the value of M10 in the first column and the start
date of the 10 most active window in the period in the second column
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

M10 is calculated by identifying the 10-hour window with the highest activity
in the given period.
}
\details{
The function calculates the averages of a quantitative variable
(typically the representative of the activity in the data set) in an interval
that comprises each observation until the end of a 10-hour window, moving on
to the next observation with the same approach. Note that the accuracy of the
M10 is related to the validity of the given data range.

The amount and date of the result will vary according to the method chosen.
Although the first approaches in actigraphy used the M10 for the entire
period (Witting et al., 1990), other approaches were studied later and
analyzes of these studies (Goncalves et al., 2015) show the importance of
checking the results for each day and for the mean profile.
}
\examples{
#Running for 1000 random observations (mean profile)
first_date <- as.POSIXct('2015-01-01')
last_date <- as.POSIXct('2015-01-11')
shuffled_timestamp <- sample(seq(first_date,
                      last_date, by = "min"), 1000)
timestamp <- sort(shuffled_timestamp)
x <- runif(1000, 0, 10000)
npcra_m10(x, timestamp, method = 2) #expects a tibble 1X2

#Ordering dates and activity data to run (each day)
data <- dplyr::as_tibble(x)
data <- dplyr::mutate(data, timestamp = shuffled_timestamp)
data <- dplyr::arrange(data, timestamp)
npcra_m10(data$value, data$timestamp, method = 3) #expects a tibble 10X2

}
\references{
WITTING, W. et al. Alterations in the circadian rest-activity rhythm in aging
and Alzheimer's disease. Biological Psychiatry, v. 27, n. 6, p. 563-572,
Mar. 1990. doi: 10.1016/0006-3223(90)90523-5.

GONCALVES, Bruno da Silva Brandao et al. A fresh look at the use of
nonparametric analysis in actimetry. Sleep Medicine Reviews, v. 20,
p. 84-91, Apr. 2015. doi: 10.1016/j.smrv.2014.06.002.
}
\seealso{
Other NPCRA functions: 
\code{\link{npcra_ism}()},
\code{\link{npcra_is}()},
\code{\link{npcra_ivm}()},
\code{\link{npcra_iv}()},
\code{\link{npcra_l5_each_day}()},
\code{\link{npcra_l5_whole_period}()},
\code{\link{npcra_l5}()},
\code{\link{npcra_m10_each_day}()},
\code{\link{npcra_m10_whole_period}()},
\code{\link{npcra_ra}()}
}
\concept{NPCRA functions}
