% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npcra.R
\name{npcra_m10_each_day}
\alias{npcra_m10_each_day}
\title{Non-Parametric Function M10 (Most Active 10 Hours) for each day}
\usage{
npcra_m10_each_day(data, col_activity = "pim", timestamp = "timestamp")
}
\arguments{
\item{data}{Dataframe that contains the date column and the column that
will be used to identify the 10 most active hours.}

\item{col_activity}{String with the name of the column that will be used in the
calculation. The observations in this column must be in numeric format.}

\item{timestamp}{String with the name of the column that contains the date
and time of each observation (POSIX format).}
}
\value{
A Dataframe with the value of M10 in the first position and the start
date of the 10 most active window by day in the second position.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

Calculate the most active window of 10 hours for each different day.
}
\examples{
\dontrun{
npcra_m10_each_day(test_log)}
}
\references{
WITTING, W. et al. Alterations in the circadian rest-activity rhythm in aging
and Alzheimer's disease. Biological Psychiatry, v. 27, n. 6, p. 563-572,
Mar. 1990. doi: 10.1016/0006-3223(90)90523-5.

GONCALVES, Bruno da Silva Brandao et al. A fresh look at the use of
nonparametric analysis in actimetry. Sleep Medicine Reviews, v. 20,
p. 84-91, Apr. 2015. doi: 10.1016/j.smrv.2014.06.002.
}
\seealso{
Other NPCRA functions: 
\code{\link{ism}()},
\code{\link{is}()},
\code{\link{ivm}()},
\code{\link{iv}()},
\code{\link{l5_average_day}()},
\code{\link{l5_each_day}()},
\code{\link{npcra_l5_whole_period}()},
\code{\link{npcra_l5}()},
\code{\link{npcra_m10_average_day}()},
\code{\link{npcra_m10_whole_period}()},
\code{\link{npcra_m10}()},
\code{\link{npcra_test_args}()},
\code{\link{ra}()}
}
\concept{NPCRA functions}
