% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npcra_m10.R
\name{npcra_m10_mean_profile}
\alias{npcra_m10_mean_profile}
\title{Non-Parametric Function M10 (Most Active 10 Hours) for the Mean Profile}
\usage{
npcra_m10_mean_profile(x, timestamp)
}
\arguments{
\item{x}{Numeric vector with the activity data that will be used in the
calculation.}

\item{timestamp}{POSIXct/POSIXlt vector that contains the date and time of
each observation.}
}
\value{
A tibble of two columns, with the value of M10 in the first column
and the start time of the 10 most active period for the mean profile in
the second position (to maintain the nomenclature standard,
this column is called start_date).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Calculates the most active window of 10 hours following the mean profile,
that is, disregarding the difference in days.

For this method the
composition of all activities is considered as a single day:
the average day.
}
\details{
The M10 is the highest average activity over a 10-hour period.
The first uses of the m10 consisted of calculating the 10 most active hours
based on the hourly means (WITTING et al., 1990). As it does not have an
explicit definition, it is expected that the values would be calculated
under the complete period (see \code{npcra_m10_whole_period()}) received
and that is why that is why many authors perform the analysis
by two other approaches (GONCALVES et al., 2015): calculation for the 24h
mean profile and for each day(see \code{npcra_m10_each_day()})

The mean profile is able to reduce the differences in activity between the
days by caring only for the time that an activity value was recorded, being
interesting to identify the most active average period

The function receives two vectors, one containing the registered activity
(numerical values) and another vector containing the date of each occurrence
(POSIXct / POSIXlt), so it is possible to map the activity and dates by the
indexes of the vectors.

m10 for the average profile starts by establishing the same date for all
observations, since it will only be necessary to check by time, and then
reordering them by the time (therefore, it is not necessary that the
observations were passed as a parameter in an orderly manner for the
optimal algorithm).

Then begins the calculation of the averages of the 10-hour periods
and the search for the largest one. It is important to note that records
whose start time is later than 2 pm include in their 10 hour period in
addition to the later hours, include records prior to 2 pm while within
the 10 hour window. For example, a record that starts at 14:30:00 will
end the 10 hour period at 00:30

At the end of the calculation, the highest average and its respective time
is captured
}
\examples{
#Using the test_log data from the package
npcra_m10_mean_profile(test_log$pim, test_log$timestamp)

#Running for 1000 random observations
first_date <- as.POSIXct('2015-01-01')
last_date <- as.POSIXct('2015-01-06')
shuffled_timestamp <- sample(seq(first_date,
                                 last_date, by = "min"), 1000)
timestamp <- sort(shuffled_timestamp)
x <- runif(1000, 0, 10000)
npcra_m10_mean_profile(x, timestamp)

}
\references{
WITTING, W. et al. Alterations in the circadian rest-activity rhythm in aging
and Alzheimer's disease. Biological Psychiatry, v. 27, n. 6, p. 563-572,
Mar. 1990. doi: 10.1016/0006-3223(90)90523-5.

GONCALVES, Bruno da Silva Brandao et al. A fresh look at the use of
nonparametric analysis in actimetry. Sleep Medicine Reviews, v. 20,
p. 84-91, Apr. 2015. doi: 10.1016/j.smrv.2014.06.002.
}
