% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npcra.R
\name{npcra_m10_whole_period}
\alias{npcra_m10_whole_period}
\title{Non-Parametric Function M10 (Most Active 10 Hours) for the entire period}
\usage{
npcra_m10_whole_period(
  data,
  col_name = "pim",
  timestamp = "timestamp",
  fast = TRUE
)
}
\arguments{
\item{data}{A \code{data.frame} \code{object}}

\item{col_name}{A string with the name of the column that will be used in the
calculation.}

\item{timestamp}{A string with the name of the column that contains the
dataframe dates.}

\item{fast}{True: It makes the M10 quick calculation, prone to errors; False:
Scans completely and returns the correct M10 value, but more slowly}
}
\value{
a Dataframe with the value of M10 in the first position and the start
date of the 10 most active window in the period in the second position.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

"M10 represents activity during the most active period of the day. This
measure may be influenced by daytime napping" - WITTING, W. (1990) This
calculation method considers the 10-hour window most active for all data.
}
\examples{
\dontrun{
npcra_m10_whole_period(test_log, fast = FALSE)}
}
\seealso{
Other NPCRA functions: 
\code{\link{npcra_is}()},
\code{\link{npcra_iv}()},
\code{\link{npcra_l5_whole_period}()},
\code{\link{npcra_l5}()},
\code{\link{npcra_m10}()},
\code{\link{npcra_test_args}()}
}
\concept{NPCRA functions}
