% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npcra.R
\name{npcra_ra}
\alias{npcra_ra}
\title{Non-Parametric Function RA (Relative Amplitude)}
\usage{
npcra_ra(data, col_activity = "pim", timestamp = "timestamp", method = 1)
}
\arguments{
\item{data}{Dataframe that contains the date column and the column with the
activity values}

\item{col_activity}{String with the name of the column that will be used in the
calculation. The observations in this column must be in numeric format.}

\item{timestamp}{String with the name of the column that contains the date
and time of each observation (POSIX format).}
}
\value{
For methods 1 and 2 returns the numerical value that represents the
relative amplitude. For method equals 3 returns a tibble with the values of
RA in the first column and the date in the second column.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

Calculate the amplitude of the rest-activity rhythm. the result is based on
the difference between the 10 most active hours (M10) and the 5 least active hours (L5).
}
\examples{
\dontrun{
npcra_ra(test_log, "pim", "timestamp", 3)}
}
\references{
WITTING, W. et al. Alterations in the circadian rest-activity rhythm in aging
and Alzheimer's disease. Biological Psychiatry, v. 27, n. 6, p. 563-572,
Mar. 1990. doi: 10.1016/0006-3223(90)90523-5.

GONCALVES, Bruno da Silva Brandao et al. A fresh look at the use of
nonparametric analysis in actimetry. Sleep Medicine Reviews, v. 20,
p. 84-91, Apr. 2015. doi: 10.1016/j.smrv.2014.06.002.
}
\seealso{
Other NPCRA functions: 
\code{\link{npcra_is}()},
\code{\link{npcra_iv}()},
\code{\link{npcra_l5_average_day}()},
\code{\link{npcra_l5_each_day}()},
\code{\link{npcra_l5_whole_period}()},
\code{\link{npcra_l5}()},
\code{\link{npcra_m10_average_day}()},
\code{\link{npcra_m10_each_day}()},
\code{\link{npcra_m10_whole_period}()},
\code{\link{npcra_m10}()},
\code{\link{npcra_test_args}()}
}
\concept{NPCRA functions}
