% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npcra.R
\name{npcra_test_args}
\alias{npcra_test_args}
\title{Testing the arguments of nonparametric functions}
\usage{
npcra_test_args(data, col_activity, timestamp, method)
}
\arguments{
\item{data}{Dataframe that contains the date column and the column that
will be used to identify the period of activity.}

\item{col_activity}{String with the name of the column that will be used in the
calculation. The observations in this column must be in numeric format.}

\item{timestamp}{String with the name of the column that contains the date
and time of each observation (POSIX format).}

\item{method}{An integer that represents one of the three common methods for
calculating and analyzing M10:

1 = whole period,

2 = average day,

3 = each day.}
}
\value{
If no test stops, TRUE will be returned.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

Performs the verification of the types of expected values as arguments to the
npcra functions.
}
\examples{
\dontrun{
npcra_test_args(test_log, "pim", "timestamp", 1)}
}
\seealso{
Other NPCRA functions: 
\code{\link{npcra_ism}()},
\code{\link{npcra_is}()},
\code{\link{npcra_ivm}()},
\code{\link{npcra_iv}()},
\code{\link{npcra_l5_average_day}()},
\code{\link{npcra_l5_each_day}()},
\code{\link{npcra_l5_whole_period}()},
\code{\link{npcra_l5}()},
\code{\link{npcra_m10_average_day}()},
\code{\link{npcra_m10_each_day}()},
\code{\link{npcra_m10_whole_period}()},
\code{\link{npcra_m10}()},
\code{\link{npcra_ra}()}
}
\concept{NPCRA functions}
