% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodogram.R
\name{periodogram}
\alias{periodogram}
\title{Compute the chi square periodogram}
\usage{
periodogram(data, breaks = "minutes", p_min = 1, p_max = 4600, step = 1)
}
\arguments{
\item{data}{An \code{\link[xts:xts]{xts}} object with a numeric vector. If the
\code{\link[xts:xts]{xts}}` have more than 1 vector, the function will use only
the first one.}

\item{breaks}{A string representing at which interval the timestamp data will
be separated to build the periodogram (seconds, minutes or hours).
(default: "minutes").}

\item{p_min}{An integer number representing the minimum period p to calculate
the test and add to the periodogram (same unit as breaks). (default: \code{1}).}

\item{p_max}{An integer number representing the maximum period p to calculate
the test and add to the periodogram (same unit as \code{breaks}). (default:
\code{4600}).}

\item{step}{An integer number representing the range of values that will be
skipped between calculating one test and the next. (default: \code{1}).}
}
\value{
A \code{\link[=list]{list}} object with the values and plots of Ap,
Qp and normalized Qp, as well as the peak and peak period.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{periodogram()} calculates the measures Ap, Qp and Qp-normalized of the
chi-square periodogram for an actigraphy dataset.

The Chi Square Periodogram is a technique to identify periodic patterns in a
time series, being widely used in chronobiology to identify the sleep-wake
cycle and its reconciliation with the 24-hour cycle. This periodogram was
proposed by Sokolove and Bushell (1978) as an adaptation of Enright's
periodogram (1965), adding the peak significance test to the method.
}
\details{
Sokolove and Bushell's periodogram assumes that the delivered data is time
equidistant - there is data for the entire time unit delivered in the breaks
parameter. If this does not occur, the function will disregard these missing
values in the middle of the time series while calculating the periodogram. If
there is more than one data for the breaks unit (for example, when the data
was captured per minute and the user wants to calculate the periodogram per
hour), the average of all values in this interval is considered as the
activity point of the periodogram.

With the formatted data, we calculate the measure Ap for each test period p,
between the minimum determined by p_min and the maximum p_max, at a step
determined by step parameter. The Ap measure of the Enright periodogram is
the standard deviation of the column means of a Buys-Ballot table. This
table has p columns and m rows, where m is a number that maximizes the
amount of values that a time series of N values can have represented in a
p columns table (thus, m is N/p rounded down). m can be seen as a cycle, and
the greater the similarity between the cycles and the difference between the
columns, more intense the standard deviation Ap.

Plotting the Ap values allows you to identify the behavior of the standard
deviations of the averages of the columns of the tables created for
different P's, and the higher the standard deviation, the more it will
tend to a peak. As these values are susceptible to high and instantaneous
fluctuations, Sokolove and Bushell proposed adding a peak significance test,
reducing the intensity of peaks by weighting the total variance of the data
and the period in which it appears. In this process, peak moments will be
highlighted among its neighbors, and smaller changes will lose prominence,
tending to a constant variation from one period to another.

The significance test of the peaks given by Ap leads to a chi-square
distribution for each p, with the test for a given p being consistent with
a distribution of P-1 degrees of freedom. The formula for calculating the
test is:

\deqn{Qp = P * m * Ap^2 / Var(X)}

Where:

\eqn{P} = the period of this test.
\eqn{m} = the total rows of the Buys-Ballot table.
\eqn{Ap} = the standard deviation of the averages from the Buys=Ballot table
of this test.
\eqn{Var(X)}= the variance of the activity data.

It is also common to use a standardized version of the test, which consists
of performing the calculation:

\deqn{Qp = Ap^2 / Var(X)}

Which results in a number between 0 and 1.
}
\examples{
data <- xts::as.xts(x = rep(seq(1, 4), times = 6),
                    order.by = seq(as.POSIXct("2020-01-01"),
                                   as.POSIXct("2020-01-01 23:00:00"),
                                   by = "hours"))
periodogram <- periodogram(data, breaks = "hours", p_max = 12)

periodogram$peak$p
#> [1] 4 # Expected
periodogram$peak$normalized_qp
#> [1] 0.9583333 # Expected
length(periodogram$qp)
#> [1] 12 # Expected

## Using interactive plots

if (interactive() &&
    requireNamespace("plotly", quietly = TRUE)) {
    plotly::ggplotly(periodogram$qp_plot)
}
}
\references{
Enright, J. T. (1965). The search for rhythmicity in biological time-series.
\emph{Journal of Theoretical Biology}, \emph{8}(3), 426–468.
\doi{10.1016/0022-5193(65)90021-4}.

Sokolove, P. G., & Bushell, W. N. (1978). The chi square periodogram: its
utility for analysis of circadian rhythms. \emph{Journal of Theoretical Biology},
\emph{72}(1), 131–160. \doi{10.1016/0022-5193(78)90022-x}.
}
\concept{analysis functions}
