% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodogram.R
\name{periodogram}
\alias{periodogram}
\title{Compute Sokolove & Bushell's \eqn{\chi^{2}}{chi square} periodogram}
\usage{
periodogram(
  data,
  col,
  p_unit = "minutes",
  p_min = 1000,
  p_max = 2500,
  p_step = 1,
  alpha = 0.05,
  print = TRUE
)
}
\arguments{
\item{data}{A \code{\link[tsibble:tsibble]{tsibble}} object.}

\item{col}{A string indicating which column of \code{data} to use.}

\item{p_unit}{(optional) a string indicating at which time unit the index
must be aggregated. By aggregating the index, this will change the time
series interval and, consequently, its \code{p} periods. Valid values are:
\verb{“seconds”}, \verb{“minutes”}, \verb{“hours”}, \verb{“days”}, \verb{“weeks”}, \verb{“months”},
\verb{“quarters”}, and \verb{“years”}) (default: \code{"minutes"}).}

\item{p_min}{(optional) an integer number indicating the minimum period
(\eqn{p}), with the same unit as \code{p_unit}, to compute the test (e.g., if
\code{p_unit = "minutes"}, \code{p_min = 1} means a period of 1 minute) (default:
\code{1000}).}

\item{p_max}{(optional) an integer number indicating the maximum period
(\eqn{p}), with the same unit as \code{p_unit}, to compute the test (default:
\code{2500}).}

\item{p_step}{(optional) an integer number indicating the range of values
that must be skipped between computing one test and the next (e.g., when
\code{p_min == 1}, \code{p_max == 7}, and \code{p_step == 2}, the test periods will be
\code{1}, \code{3}, \code{5}, and \code{7}) (default: \code{1}).}

\item{alpha}{(optional) a number, from \code{0} to \code{1}, indicating the significant
level (\eqn{\alpha}{alpha}) required for the peak significance test
(default: \code{0.05}).}

\item{print}{(optional) a \code{\link[=logical]{logical}} value indicating if the
function must print the \eqn{Q_{p}}{Qp} plot (default: \code{TRUE}).}
}
\value{
A \code{\link[=list]{list}} object with the following elements:
\itemize{
\item \code{p_unit}: a string indicating the time unit in which the data was
aggregated.
\item \code{p_seq}: a \code{\link[=numeric]{numeric}} object with the the sequence of the
tested periods.
\item \code{alpha}: a number indicating the significant level used.
\item \code{a_p}: a \code{\link[=numeric]{numeric}} object with the root mean square amplitude
(\eqn{A_{p}}{Ap}) for each period.
\item \code{a_p_plot}: a \code{\link[ggplot2:ggplot]{ggplot}} object with a line chart
showing \code{a_p} (y) by \code{p_seq} (x).
\item \code{q_p}: a \code{\link[=numeric]{numeric}} object with the peak significant test
(\eqn{Q_{p}}{Qp}) for each period .
\item \code{q_p_critical}: a \code{\link[=numeric]{numeric}} object with the
\eqn{\chi^{2}}{chi square} critical values for each \code{q_p}, based on the
\code{alpha} parameter.
\item \code{q_p_pvalue}: a \code{\link[=numeric]{numeric}} object with the p-value for each
\code{q_p}, based on the \code{alpha} parameter.
\item \code{q_p_peaks}: a \code{\link[tibble:tibble]{tibble}} object listing each period that
peaked above the critical value along with its \code{q_p}, \code{q_p_critical},
relative \code{q_p} (\code{q_p_rel = q_p_critical - q_p}), and \code{q_p_pvalue}.
\item \code{q_p_plot}: a \code{\link[ggplot2:ggplot]{ggplot}} object with a line chart
showing \code{q_p} (y1) and \code{q_p_critical} (y2) by \code{p_seq} (x).
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{periodogram()} returns the Sokolove & Bushell's \eqn{\chi^{2}}{chi square}
periodogram plot and measures for a \code{\link[tsibble:tsibble]{tsibble}} object.

The \eqn{\chi^{2}}{chi square} periodogram is a technique to identify
periodic patterns in a time series, being used in chronobiology to identify
the presence/absence of circadian periodicities in rest-activity data. It was
proposed by Sokolove and Bushell (1978) as an adaptation of Enright's
periodogram (1965), adding the peak significance test to the procedure.
}
\details{
\subsection{Irregular intervals}{

Sokolove and Bushell's periodogram assumes that the time series is regular,
i.e., that it has regular intervals/epochs. \code{periodogram()} will still work
with slightly irregular time series, but only if the latter have at least a
90\% prevalence of the same periodicity. Warning messages will be issued
telling the user if any data irregularities were found.
}

\subsection{Data aggregation}{

If there's more than one data for each \code{p_unit} (e.g., when the data was
recorded per minute and the user wants to compute the periodogram per hour),
\code{periodogram()} will aggregate the values by averaging them or by assigning
the most frequent value (mode) (for non-numeric or single integer variables).
}

\subsection{Missing values}{

\code{periodogram()} will work even with missing values. As is the case for
any analysis with missing data, the results may diverge. You may want to
interpolate these data points.

There are few articles that deals with interpolation in actigraphy. Tonon et
al. (2022) recommends not using interpolation (i.e., maintain \code{NA} values)
whenever is possible. The same authors also recommends using the
\link[=?na_weekly_mean]{weekly mean method} of interpolation when the parameters
cannot be computed in the presence of \code{NA} values.
}
}
\section{Guidelines}{


Enright (1965) and Sokolove & Bushell (1978) guidelines for the
\eqn{\chi^{2}}{chi square} periodogram computation are as follows.
\subsection{Notes}{
\itemize{
\item If you are visualizing this documentation in plain text (\code{ASCII}), you may
have some trouble understanding the equations. If you want a better viewer,
you can see this documentation on the package
\href{https://gipso.github.io/actverse/reference/}{website}.
}
}

\subsection{\eqn{A_{p}}{Ap}: The statistic adopted to express "importance"}{

\code{periodogram()} compute the \eqn{A_{p}}{Ap} stat for each test period
\eqn{p}, between the minimum determined by \code{p_min} argument and the maximum
\code{p_max} argument, at a step determined by \code{p_step} argument (e.g., when
\code{p_min == 1}, \code{p_max == 5}, and \code{p_step == 1}, the test periods will be
\code{1}, \code{2}, \code{3}, \code{4} and \code{5}).

The \eqn{A_{p}}{Ap} measure of Enright's periodogram is the standard
deviation of column means of a Buys-Ballot table, or, as Enright puts it,
"the root mean square \strong{amplitude}". This Buys-Ballot table has \eqn{P}
columns and \eqn{m} rows, where \eqn{m} is a number that maximizes the amount
of values that a time series of \eqn{N} values can have represented in a
\eqn{P} columns table (thus, \eqn{m} is \eqn{N / p} rounded down). \eqn{m}
can be seen as a cycle, and the greater the similarity between the cycles and
the difference between the columns, more intense will be the standard
deviation (\eqn{A_{p}}{Ap}).
\itemize{
\item Buys-Ballot table for an integral period \eqn{p}:
}\preformatted{                               P (count)
          |------------------------------------------------|     -
                                                                 |
Row 1          X_1          X_2          ...          X_P        |
Row 2         X_P+1        X_P+2         ...          X_2P       | m (count)
...            ...          ...          ...          ...        |
Row m      X_P(m-1)+1   X_P(m-1)+2       ...          X_Pm       |
          --------------------------------------------------     -
Totals        U_P,1        U_P,2         ...          U_P,P
Averages      Y_P,1        Y_P,2         ...          Y_P,P
}

As the table above shows, \eqn{P} is the number of columns in the array
(matrix/table), while \eqn{m} is the number of rows. When \eqn{p} (test
period) is an integer, \eqn{p = P}. That way, \eqn{A_{p}}{Ap} can be computed
as:

\strong{\deqn{A_{p} = \sqrt{\frac{\sum^{P}_{h = 1} (Y_{p, h} -
\overline{Y}_{p})^{2}}{P}}}{Ap = ((sum^{P}_{h = 1} (Y_p,h - MeanY_p)^{2}) /
P)^{1/2}}}

In which:

\strong{\deqn{\overline{Y}_{p} = \frac{\sum^{P}_{h = 1} Y_{p, h}}{P}}{MeanY_p =
(sum^{P}_{h = 1} Y_p,h) / P}}
}

\subsection{\eqn{Q_{p}}{Qp}: Sokolove & Bushell's peak significance test}{

Plotting the \eqn{A_{p}}{Ap} values allows you to identify the magnitude of
the standard deviations for different \eqn{p}s. The higher the standard
deviation the more it will tend to a peak. As these values are susceptible to
high and instantaneous fluctuations, Sokolove & Bushell proposed adding a
peak significance test, reducing the intensity of peaks by weighting the
total variance of the data and the period in which it appears. In this
process, peak moments will be highlighted among its neighbors, and smaller
changes will lose prominence, tending to a constant variation from one period
to another.

The significance test of the peaks given by \eqn{A_{p}}{Ap} leads to a
\eqn{\chi^{2}}{chi square} distribution for each \eqn{p}, with the test for a
given \eqn{p} being consistent with a distribution of \eqn{P - 1} degrees of
freedom.

The formula for calculating the test is:

\strong{\deqn{Q_{p} = \frac{P \times Ap^{2}}{\sigma^{2}_{\overline{X}}}}{Qp =
P * m * Ap^2 / Var(X_line)}}

In which:

\strong{\deqn{\sigma^{2}_{\overline{X}} = \frac{\sigma^{2}_{X}}{m}}{
Var(X_line) = Var(X) / m}}

Where:
\itemize{
\item \eqn{P} = number of columns of the test Buys-Ballot table.
\item \eqn{m} = number of rows of the test Buys-Ballot table.
\item \eqn{A_{p}}{Ap} = standard deviation of the column averages from the test
Buys-Ballot table.
\item \eqn{\sigma^{2}_{\overline{X}}}{Var(X_line)}= variance of the test data
(\eqn{X}).
}
}
}

\examples{
data <- dplyr::tibble(
    index = seq(as.POSIXct("2020-01-01"),
                as.POSIXct("2020-01-02 05:59:59"),
                by = "min"),
    x = rep(seq(1, 60), times = 30))
data <- tsibble::tsibble(data, index = index)

per <- periodogram(data, "x", p_unit = "minutes", p_min = 1,
                   p_max = 350, p_step = 1, alpha = 0.05,
                   print = TRUE)

## Using interactive plots

if (interactive() &&
    requireNamespace("plotly", quietly = TRUE)) {
    plotly::ggplotly(per$q_p_plot)
}
}
\references{
Enright, J. T. (1965). The search for rhythmicity in biological time-series.
\emph{Journal of Theoretical Biology}, \emph{8}(3), 426-468.
\doi{10.1016/0022-5193(65)90021-4}.

Sokolove, P. G., & Bushell, W. N. (1978). The chi square periodogram: its
utility for analysis of circadian rhythms. \emph{Journal of Theoretical Biology},
\emph{72}(1), 131-160. \doi{10.1016/0022-5193(78)90022-x}.

Tonon, A. C. et al. (2022). Handling missing data in rest-activity
time series measured by actimetry. \emph{Chronobiology International}.
\doi{10.1080/07420528.2022.2051714}.
}
\seealso{
Other period analysis functions: 
\code{\link{spectrogram}()}
}
\concept{period analysis functions}
