% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_acttrust.R
\name{read_acttrust}
\alias{read_acttrust}
\title{Read, tidy, and validate an ActTrust file}
\usage{
read_acttrust(
  file = tcltk::tk_choose.files(mult = FALSE),
  tz = "UTC",
  regularize = TRUE
)
}
\arguments{
\item{file}{(optional) A string with the file path for the ActTrust data. If
not assigned, a dialog window will open allowing the user to browse and
select a file (default: \code{\link[tcltk:tk_choose.files]{tk_choose.files()}}).}

\item{tz}{(optional) A string that specifies which time zone to parse the
dates/time with. The string must be a time zone that is recognized by the
user's OS. For more information see \code{\link[=OlsonNames]{?timezone}}
(default: \code{"UTC"}).}

\item{regularize}{(optional) A \code{\link[=logical]{logical}} value indicating if the
function must correct irregular intervals (\strong{strongly recommended}). See
more about it in the Details section (default: \code{TRUE}).}
}
\value{
A \link[tsibble:tsibble]{tsibble} object. The data structure can be
found in \code{\link[=acttrust]{?acttrust}}.
}
\description{
\code{read_acttrust()} allows you to read, tidy, and validate an ActTrust file in
a consistent and easy manner. You can see the output data structure in
\code{\link[=acttrust]{?acttrust}}.

ActTrust is a trademark of
\href{https://condorinst.com/en/}{Condor Instruments Ltda}.
}
\details{
\subsection{\code{regularize} parameter}{

ActTrust data files may have uneven epochs or intervals due to small drifts
in the device's internal clock. These irregularities can affect analyses.
Setting \code{regularize = TRUE} in \code{read_acttrust()} will detect and correct
such issues by regularizing the time intervals.

Regularization is performed only if a clear epoch or periodicity is found.
During this process, values within each epoch are aggregated: numeric
variables are averaged, and categorical or integer variables are assigned
their most frequent value (mode).

Any gaps in the time series are filled with \code{NA}, and the corresponding
\code{state} is set to \code{9}.
}

\subsection{Offwrist data}{

\code{read_acttrust()} will convert all offwrist data (where \code{state == 4}) to
missing values (\code{NA}). These data points will remain classified as offwrist
in the \code{state} variable.
}
}
\examples{
get_raw_data("acttrust.txt") |> read_acttrust()
}
\seealso{
Other read/write functions: 
\code{\link{write_acttrust}()}
}
\concept{read/write functions}
