% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_acttrust.R
\name{read_acttrust}
\alias{read_acttrust}
\title{Read and tidy an ActTrust file}
\usage{
read_acttrust(file = file.choose(), tz = "America/Sao_Paulo")
}
\arguments{
\item{file}{A string with the file path for the ActTrust dataset. If not
assigned, a dialog window will open allowing the user to browse and select
a file.}

\item{tz}{A string that specifies which time zone to parse the dates/time
with. The string must be a time zone that is recognized by the user's OS.
For more information, see \code{\link[lubridate:ymd_hms]{?lubridate::ymd_hms}}.
(default: \code{"America/Sao_Paulo"}).}
}
\value{
An \link[xts:xts]{xts} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{read_acttrust()} reads and tidy an ActTrust file.

\code{read_acttrust()} was created just for convenience. If this function doesn't
work for your file, we recommend using the
\href{https://readr.tidyverse.org/}{readr} package to load it to R.

ActTrust is a trademark of
\href{https://www.condorinst.com.br/}{Condor Instruments}.
}
\examples{
\dontrun{
read_acttrust()
}
}
\seealso{
Other utility functions: 
\code{\link{find_epoch}()}
}
\concept{utility functions}
