% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrogram.R
\name{spectrogram}
\alias{spectrogram}
\title{Compute a spectrogram based on Sokolove & Bushell's periodogram}
\usage{
spectrogram(
  data,
  breaks = "minutes",
  p_min = 1000,
  p_max = 2500,
  p_step = 1,
  int = "days",
  n_int = 7,
  epoch_step = 720,
  alpha = 0.8,
  print = TRUE
)
}
\arguments{
\item{data}{An \code{\link[xts:xts]{xts}} object with a numeric vector. If the
\code{\link[xts:xts]{xts}}` have more than 1 vector, the function will use only
the first one.}

\item{breaks}{A string indicating at which interval the index must be
aggregated. (valid values: \verb{“microseconds”}, \verb{“milliseconds”}, \verb{“seconds”},
\verb{“minutes”}, \verb{“hours”}, \verb{“days”}, \verb{“weeks”}, \verb{“months”}, \verb{“quarters”}, and
\verb{“years”}) (default: \code{"minutes"}).}

\item{p_min}{An integer number indicating the minimum period p to calculate
the test and add to the periodogram (same unit as \code{breaks}). (default:
\code{1}).}

\item{p_max}{An integer number indicating the maximum period p to calculate
the test and add to the periodogram (same unit as \code{breaks}). (default:
\code{4600}).}

\item{p_step}{An integer number indicating the range of values that will be
skipped between calculating one test and the next. (default: \code{1}).}

\item{int}{A string indicating the interval unit (default: \code{"days"}).}

\item{n_int}{An integer number indicating the amount of intervals.}

\item{epoch_step}{An integer number indicating the amount of epochs to
advance at the end of each interval.}

\item{alpha}{A number, from 0 to 1, indicating the significant level required
for the peaks (default: \code{0.9}). The spectrogram plot only shows the
significant peaks. A high \code{alpha} will produce a more detailed image.}

\item{print}{A \code{\link[=logical]{logical}} value indicating if the function
must print the spectrogram plot (default: \code{TRUE}).}
}
\value{
a \code{\link[=list]{list}} object with the following elements:
\itemize{
\item \code{periodograms}: A \code{\link[=list]{list}} object with the periodograms data.
\item \code{spectogram}: a \code{\link[ggplot2:ggplot]{ggplot}} object with a heat map plot
of the spectrogram.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{spectrogram()} compute a series of Sokolove & Bushell's \eqn{chi^{2}}{chi
square} periodograms with the purpose of visualize differences in
periodicities in a given interval of a time series.
}
\examples{
data <- xts::as.xts(x = rep(seq(1, 60), times = 30),
                    order.by = seq(as.POSIXct("2020-01-01"),
                                   as.POSIXct("2020-01-02 05:59:59"),
                                   by = "min"))
spectrogram <- spectrogram(data, breaks = "minutes", p_min = 1, p_max = 120,
                           p_step = 1, int = "hours", n_int = 2,
                           epoch_step = 59, alpha = 0.5, print = FALSE)

head(names(spectrogram$periodograms))
spectrogram$periodograms$int_1$unit
spectrogram$periodograms$int_1$alpha
spectrogram$periodograms$int_1$interval
head(spectrogram$periodograms$int_1$periods)
head(spectrogram$periodograms$int_1$a_p)
head(spectrogram$periodograms$int_1$q_p)
head(spectrogram$periodograms$int_1$q_p_alpha)
head(spectrogram$periodograms$int_1$q_p_rel)
head(spectrogram$periodograms$int_1$q_p_pvalue)
head(spectrogram$periodograms$int_1$q_p_rel)
spectrogram$periodograms$int_1$q_p_peaks

spectrogram$spectrogram

## Using interactive plots

if (interactive() &&
    requireNamespace("plotly", quietly = TRUE)) {
    plotly::ggplotly(spectrogram$spectrogram)
}
}
\seealso{
Other analysis functions: 
\code{\link{periodogram}()}
}
\concept{analysis functions}
