% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_log.R
\docType{data}
\name{test_log}
\alias{test_log}
\title{An actigraphy record}
\format{
A tibble with 24 columns and 10074 rows:

\describe{
\item{timestamp}{
Log timestamp.
\cr \cr
R class: \code{POSIXct}.}

\item{ms}{
Log millisecond.
\cr \cr
R class: \code{numeric}.}

\item{x_axis}{
X axis signal.
\cr \cr
R class: \code{numeric}.}

\item{y_axis}{
Y axis signal.
\cr \cr
R class: \code{numeric}.}

\item{z_axis}{
Z axis signal.
\cr \cr
R class: \code{numeric}.}

\item{pim}{
Wrist activity in Proportional Integral Mode (PIM).
\cr \cr
R class: \code{numeric}.}

\item{pim_n}{
Wrist activity in normalized PIM value per second (PIM divided by epoch
seconds).
\cr \cr
R class: \code{numeric}.}

\item{tat}{
Wrist activity in Time Above Threshold (TAT) mode.
\cr \cr
R class: \code{numeric}.}

\item{tat_n}{
Wrist activity in normalized TAT value per second (TAT divided by epoch
seconds).
\cr \cr
R class: \code{numeric}.}

\item{zcm}{
Wrist activity in Zero Crossing Mode (\code{ZCM}).
\cr \cr
R class: \code{numeric}.}

\item{zcm_n}{
Wrist activity in normalized ZCM value per second (ZCM divided by epoch
seconds).
\cr \cr
R class: \code{numeric}.}

\item{orientation}{
Device orientation.
\cr \cr
R class: \code{numeric}.}

\item{body_temperature}{
Body temperature in degrees \code{Celsius} (wrist skin temperature).
\cr \cr
R class: \code{numeric}.}

\item{external_temperature}{
External temperature in degrees \code{Celsius}.
\cr \cr
R class: \code{numeric}.}

\item{light}{
Light intensity in \code{lux}.
\cr \cr
R class: \code{numeric}.}

\item{ambient_light}{
Ambient light intensity in \verb{microwatts/square meter}.
\cr \cr
R class: \code{numeric}.}

\item{red_light}{
Red light intensity in \verb{microwatts/square centimeter}.
\cr \cr
R class: \code{numeric}.}

\item{green_light}{
Green light intensity in \verb{microwatts/square centimeter}.
\cr \cr
R class: \code{numeric}.}

\item{blue_light}{
Blue light intensity in \verb{microwatts/square centimeter}.
\cr \cr
R class: \code{numeric}.}

\item{ir_light}{
Infrared light intensity in \verb{microwatts/square centimeter}.
\cr \cr
R class: \code{numeric}.}

\item{uva_light}{
Ultraviolet A light intensity in \verb{microwatts/square centimeter}.
\cr \cr
R class: \code{numeric}.}

\item{uvb_light}{
Ultraviolet B light intensity in \verb{microwatts/square centimeter}.
\cr \cr
R class: \code{numeric}.}

\item{event}{
A \code{logical} value indicating whether there was an event in the sampling
period.
\cr \cr
R class: \code{logical}.}

\item{state}{
A \code{factor} value indicating the subject status: \code{"Awake"}, \code{"Resting"},
\code{"Sleeping"}, \code{"OffWrist"} (device removed) or editable states.
\cr \cr
R class: \code{factor}.}
}
}
\source{
Prepared by Daniel Vartanian (package's author) (personal log).
}
\usage{
test_log
}
\description{
An actigraphy record, free for all kinds of use, personal made, and granted
by one of the authors of the package.

\strong{Device}: \href{https://www.condorinst.com.br/}{Condor Instruments} ActTrust 1
}
\references{
Condor Instruments. (2019). \emph{ActTrust - user manual}. [Sao Paulo, Brazil]:
[s.n.]. Model: AT0503. ActStudio version: 1.0.13.
}
\concept{datasets}
\keyword{datasets}
