% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_acttrust.R
\name{write_acttrust}
\alias{write_acttrust}
\title{Write a \code{tsibble} to a readable ActTrust file}
\usage{
write_acttrust(data, file, delim = ";", header = NULL)
}
\arguments{
\item{data}{A \code{\link[tsibble:tsibble]{tsibble}} object with a
\code{\link[=as.POSIXct]{POSIXt}} index. The ActTrust software only deals with this
type of index.}

\item{file}{A string with a file path to write to.}

\item{delim}{(optional) A string indicating the delimiter that must be used
to separate values. Valid delimiters are: \code{";"} and \code{"\t"} (default:
\code{";"}).}

\item{header}{(optional) A string indicating the path to a file (usually
the raw data file) with the ActTrust header. This is not mandatory,
the ActTrust software can read files without it (default: \code{NULL}).}
}
\value{
An invisible \code{NULL}. This function is used for its side effect.
}
\description{
\code{write_acttrust()} allows you to adapt and write a
\code{\link[tsibble:tsibble]{tsibble}} object to a readable ActTrust file.

Note that your data must conform to a predefined structure. See
\code{\link[=acttrust]{?acttrust}} to learn more.

ActTrust is a trademark of
\href{https://condorinst.com/en/}{Condor Instruments Ltda}.
}
\details{
\subsection{\code{NA} values}{

\code{write_acttrust()} will transform any \code{NA} value to \code{0}. This is because
the ActTrust software only deals with numeric values.
}
}
\examples{
library(readr)

acttrust

file <- tempfile()

write_acttrust(
  data = acttrust,
  file = file,
  header = get_raw_data("acttrust.txt")
)

readr::read_lines(file, n_max = 30)
}
\seealso{
Other read/write functions: 
\code{\link{read_acttrust}()}
}
\concept{read/write functions}
